% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_stats}
\alias{add_stats}
\title{Add decision statistics to data (containing counts of a 2x2 contingency table)}
\usage{
add_stats(
  data,
  sens.w = 0.5,
  cost.each = NULL,
  cost.outcomes = list(hi = 0, fa = 1, mi = 1, cr = 0)
)
}
\arguments{
\item{data}{A data frame with (integer) values named \code{"hi"}, \code{"mi"}, \code{"fa"}, and \code{"cr"}.}

\item{sens.w}{numeric. Sensitivity weight (for computing weighted accuracy, \code{wacc}).}

\item{cost.each}{numeric. An optional fixed cost added to all outputs (e.g.; the cost of the cue).}

\item{cost.outcomes}{list. A list of length 4 named \code{"hi"}, \code{"mi"}, \code{"fa"}, \code{"cr"},  and
specifying the costs of a hit, miss, false alarm, and correct rejection, respectively.
E.g.; \code{cost.outcomes = listc("hi" = 0, "fa" = 10, "mi" = 20, "cr" = 0)} means that a
false alarm and miss cost 10 and 20 units, respectively, while correct decisions incur no costs.}
}
\description{
\code{add_stats} assumes \code{data} input with 2x2 frequency counts
(named \code{"hr"}, \code{"cr"}, \code{"mi}, and \code{"fa"}) and
computes various decision accuracy and cost measures.
}
\details{
Providing \code{cost.each} and \code{cost.outcomes} (as a named list) allows computing
cost information for the counts of corresponding classification decisions.
}
