% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applytree_function.R
\name{apply.tree}
\alias{apply.tree}
\title{Applies a fast and frugal tree to a dataset.}
\usage{
apply.tree(data, formula, tree.definitions)
}
\arguments{
\item{data}{dataframe. A model training dataset. An m x n dataframe containing n cue values for each of the m exemplars.}

\item{formula}{A formula}

\item{tree.definitions}{dataframe. A dataframe defining the tree structure(s)}
}
\value{
A list of length 4.
}
\description{
Applies a fast and frugal tree to a dataset.
}

