% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.data.frame,cue-method}
\alias{as.data.frame}
\alias{as.data.frame,cue-method}
\title{Converts \code{cue} to a \code{data.frame}}
\usage{
\S4method{as.data.frame}{cue}(x, row.names = NULL, optional = NULL, ...)
}
\arguments{
\item{x}{cue}

\item{row.names}{not used.}

\item{optional}{not used}

\item{...}{not used}
}
\value{
data.frame
}
\description{
Converts \code{cue} to a \code{data.frame}
}
\seealso{
\code{\link{cue}}
}

