% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.data.frame,fftree-method}
\alias{as.data.frame}
\alias{as.data.frame,fftree-method}
\title{Converts \code{fftree} to a \code{data.frame}}
\usage{
\S4method{as.data.frame}{fftree}(x, row.names = NULL, optional = NULL, ...)
}
\arguments{
\item{x}{Fast and frugal tree object}

\item{row.names}{not used}

\item{optional}{string Can either be "treedevelopment" or "cueefficiency". Default is "treedevelopment"}

\item{...}{not used}
}
\value{
data.frame
}
\description{
Converts \code{fftree} to a \code{data.frame}
}
\seealso{
\code{\link{Sdt},\link{fftree}}
}

