% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_function.R
\name{auc}
\alias{auc}
\title{Calculates AUC (Area under the Curve) using trapezoidal approximation}
\usage{
auc(hr.v, far.v)
}
\arguments{
\item{hr.v}{a vector of hit rates}

\item{far.v}{A vector of false alarm rates}
}
\description{
Calculates AUC (Area under the Curve) using trapezoidal approximation
}
\examples{

# Calculate the AUC for a vector of hit rates and false alarm rates

auc(hr.v = c(.1, .3, .5, .7), far.v = c(.05, .1, .15, .3))


}

