% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{blood}
\alias{blood}
\title{Blood donation data}
\format{
A data frame containing 748 rows and 5 columns.

\describe{
  \item{recency}{Months since last donation}
  \item{frequency}{Total number of donations}
  \item{total}{Total blood donated (in c.c.)}
  \item{time}{Months since first donation}
  \item{donation.crit}{\emph{Criterion}: Did the person donate blood (in March 2007)?

  Values: \code{0}/no vs. \code{1}/yes (76.2\% vs.\ 23.8\%).}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Blood+Transfusion+Service+Center}



Original owner and donor:

Prof. I-Cheng Yeh

Department of Information Management

Chung-Hua University
}
\usage{
blood
}
\description{
Data taken from the Blood Transfusion Service Center in Hsin-Chu City in Taiwan
}
\seealso{
Other datasets: 
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
