% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bruteforcemaximize}
\alias{bruteforcemaximize}
\alias{fftm.bruteforce.maximize}
\title{Maximizes a \code{\link{Sdt}} value by bruteforcing the tree permutation}
\usage{
fftm.bruteforce.maximize(tree, whatToMaximize)
}
\arguments{
\item{tree}{\code{\link{fftree}}}

\item{whatToMaximize}{string; Which parameter of the tree should be maximized? See \link{Sdt} for possible values.}
}
\value{
list with best cues
}
\description{
Takes given \code{\link{fftree}}, and permutates given \code{\link{cue}s} in all possible variants
to find a permutation of cues, which has the highest value of \code{whatToMaximize} in \code{\link{Sdt}}. Careful: This may take a LONG while!
}
\details{
This is the most inefficient method, but the most secure. For a more performant approach with similar results, use \link{maximize} instead. Everything from \link{Sdt} can be maximized.
}
\seealso{
\code{\link{Fftm}}, \code{\link{maximize}}, \code{\link{chase}}, , \code{\link{montecarlo}}
}

