% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{car}
\alias{car}
\title{Car acceptability data}
\format{
A data frame containing 1728 cars (rows) and 7 variables (columns).

\describe{
  \item{buying.price}{Numeric}
  \item{maint.price}{price of the maintenance, Factor}
  \item{doors}{number of doors, Factor}
  \item{persons}{capacity in terms of persons to carry, Numeric}
  \item{luggage}{the size of luggage boot, Numeric}
  \item{safety}{estimated safety of the car, Factor}
  \item{acceptability}{\emph{Criterion}: Category of acceptability rating (as factor).
  Values: \code{unacc}/\code{acc}(77.78\% vs. 22.2\%)}
}
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Car+Evaluation}

Original creator and donor:

Marko Bohanec and Blaz Zupan
}
\usage{
car
}
\description{
A dataset on car evaluations based on basic features, derived from a simple hierarchical decision model.
}
\details{
The criterion variable is a car's \code{acceptability} rating.


We made the following enhancements to the original data for improved usability:

 - The criterion was binarized from a factor variable with four levels (\code{unacc}/\code{acc}/\code{good}/\code{vgood} , into a factor variable with two levels (\code{unacc}/\code{acc}).

 Other than that, the data remains consistent with the original dataset.
}
\references{
Bohanec, M., Rajkovic, V. (1990): Expert system for decision making.
\emph{Sistemica}, \emph{1} (1), 145--157.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
