% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{classtableC}
\alias{classtableC}
\title{classtableC}
\usage{
classtableC(pred, crit)
}
\arguments{
\item{pred}{A binary vector of predictions}

\item{crit}{A binary vector of criterion (true) values}
}
\description{
Calculates several classification statistics from binary prediction and criterion (e.g.; truth) vectors
}
