% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stats.R
\name{comp_pred}
\alias{comp_pred}
\title{A wrapper for competing classification algorithms}
\usage{
comp_pred(
  formula,
  data.train,
  data.test = NULL,
  algorithm = NULL,
  model = NULL,
  sens.w = NULL,
  new.factors = "exclude"
)
}
\arguments{
\item{formula}{A formula (usually \code{x$formula}, for an \code{FFTrees} object \code{x}).}

\item{data.train}{A training dataset (as a data frame).}

\item{data.test}{A testing dataset (as a data frame).}

\item{algorithm}{A character string specifying an algorithm in the set:
\itemize{
  \item{\code{"lr"}: Logistic regression (using \code{\link{glm}} from \strong{stats} with \code{family = "binomial"});}
  \item{\code{"rlr"}: Regularized logistic regression (currently not supported);}
  \item{\code{"cart"}: Decision trees (using \code{rpart} from \strong{rpart});}
  \item{\code{"svm"}: Support vector machines (using \code{svm} from \strong{e1071});}
  \item{\code{"rf"}: Random forests (using \code{randomForest} from \strong{randomForest}.}
}}

\item{model}{An optional existing model (as a \code{model}), to be applied to the test data.}

\item{sens.w}{Sensitivity weight parameter (numeric, from \code{0} to \code{1}), required to compute \code{wacc}.}

\item{new.factors}{What should be done if new factor values are discovered in the test set (as a character string)?
Available options:
\itemize{
  \item{\code{"exclude"}: exclude case (i.e., remove these cases, used by default);}
  \item{\code{"base"}: predict the base rate of the criterion.}
}}
}
\description{
\code{comp_pred} provides the main wrapper for running alternative classification algorithms, such as CART (\code{rpart::rpart}),
logistic regression (\code{glm}), support vector machines (\code{svm::svm}), and random forests (\code{randomForest::randomForest}).
}
