% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CONST_SAVE}
\alias{CONST_SAVE}
\title{CONST_SAVE}
\format{\preformatted{ num 2
}}
\usage{
CONST_SAVE
}
\description{
CONST_SAVE
}
\details{
internal constant for \code{\link{permu.new}} Represents the vector-index of the SAVE flag.
}
\keyword{datasets}

