% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CONST_SKIP}
\alias{CONST_SKIP}
\title{CONST_SKIP}
\format{\preformatted{ num 1
}}
\usage{
CONST_SKIP
}
\description{
CONST_SKIP
}
\details{
internal constant for \code{\link{permu.new}}. Represents the vector-index of the SKIP flag.
}
\keyword{datasets}

