% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CONST_VAL}
\alias{CONST_VAL}
\title{CONST_VAL}
\format{\preformatted{ num 3
}}
\usage{
CONST_VAL
}
\description{
CONST_VAL
}
\details{
internal constant for \code{\link{permu.new}} Represents the vector-index of the user defined VAL-parameter.
}
\keyword{datasets}

