% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{contraceptive}
\alias{contraceptive}
\title{Contraceptive use data}
\format{
A data frame containing 1473 cases (rows) and 10 variables (columns).

\describe{
  \item{wife.age}{Numeric}
  \item{wife.edu}{Factor}
  \item{hus.ed}{Factor}
  \item{children}{Numeric}
  \item{wife.rel}{Numeric}
  \item{wife.work}{Factor}
  \item{hus.occ}{Factor}
  \item{sol}{Factor}
  \item{media}{Numeric}
  \item{cont.crit}{\emph{Criterion}: Use of a contraceptive (as logical).
  Values: \code{FALSE} vs. \code{TRUE} (42.7\% vs.\ 57.3\%).}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Contraceptive+Method+Choice}
}
\usage{
contraceptive
}
\description{
A subset of the 1987 National Indonesia Contraceptive Prevalence Survey.
}
\details{
The samples describe married women who were either not pregnant
or do not know if they were pregnant at the time of the interview.

The problem consists in predicting a woman's current contraceptive method choice
(here: binarized \code{cont.crit})
based on her demographic and socio-economic characteristics.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
