% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPartition_function.R
\name{createPartition}
\alias{createPartition}
\title{Splits a dataset into two partitions}
\usage{
createPartition(data, formula, train.p)
}
\arguments{
\item{data}{dataframe.}

\item{formula}{formula. A formula specifying the criterion}

\item{train.p}{numeric. The percentage of data used to create the training set}
}
\value{

}
\description{
Splits a dataset into two partitions
}
\examples{

 library(FFTrees) # For the heartdisease data

 heartdisease.split <- createPartition(data = heartdisease,
                                 formula = diagnosis ~.,
                                 train.p = .1)


 heartdisease.split$part1   # The training data
 heartdisease.split$part2    # The testing data





}
