% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{cue-class}
\alias{Cue}
\alias{cue}
\alias{cue-class}
\title{Class cue}
\usage{
Cue(inputvector, test = "<=", split = 0, pred = TRUE, name = "")
}
\arguments{
\item{inputvector}{Data, which should be predicted. If you just want to build a tree without measuring it's performance, set to c(0)}

\item{test}{Charactervector, which describes the binarization- operator. Possible values \code{'==', '<=', '<', '>', '>='}}

\item{split}{Numeric value; at which the prediction switches}

\item{pred}{Logical value; what the cue will predict, if the \code{test} at the particular \code{split} succeeds}

\item{name}{Charactervector; how the cue is named (just for description, without function)}
}
\value{
cue S4 object
}
\description{
Class \code{cue} is a node in a tree
with parameters to binarize the data

A \code{cue} is an S4 object, that represents a node in a tree.
Fast- and Frugal trees use binary predicitons, but data
normally is numeric. By defining the cue- parameters
the cue automatically binarizes given data by a set of parameters.
}
\seealso{
\code{\link{Fftree}}, \code{\link{cue.getAllVariants}}
}

