% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cue.getAllVariants}
\alias{cue.getAllVariants}
\title{Creates list of all possible variants of cue- configurations.}
\usage{
cue.getAllVariants(inputvector, allSplitValues = NULL, allTests = NULL,
  allPredictions = NULL, name = "Cue")
}
\arguments{
\item{inputvector}{Data, which should be predicted.}

\item{allSplitValues}{Numeric value; at which the prediction switches}

\item{allTests}{character-/ stringvector, which describes the binarization- operator. ('==', '<=', '<', etc.)}

\item{allPredictions}{Logical value; what the cue will predict, if the \code{test} at the particular \code{split} succeeds}

\item{name}{Charactervector; how the cues should be named}
}
\value{
list with S4 cue objects
}
\description{
By defining the optional parameters, these variants will be fix.
Parameters with NULL- value will be varied
}
\seealso{
\code{\link{Cue}} \code{\link{Fftree}}
}

