% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cv}
\alias{crossvalidate}
\alias{cv}
\title{Crossvalidates \code{fftree} with another data- sample}
\usage{
crossvalidate(tree, data = tree@crossvalidationdata)
}
\arguments{
\item{tree}{fftree.}

\item{data}{dataframe or list with dataframes  similar to data the tree was fitted on}
}
\value{
list with 2 items. "MAE.leafs": Errors of leafs (exit structures of the fftree) crossvalidated predictions and "mean absolute error". "MAE.overall": Errors of fftree crossvalidated predictions and "mean absolute error".
}
\description{
Crossvalidates \code{fftree} with another data- sample
}
\seealso{
\code{\link{cue}},\code{\link{fftree}}, \code{\link{isCuePrunable}}, \code{\link{isCueLazy}}, \code{\link{isCueTrivial}}
}

