% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{fft}
\alias{fft}
\title{Deprecated functions}
\usage{
fft(...)
}
\arguments{
\item{...}{arguments passed from the old functions to the new functions}
}
\description{
These functions have been renamed and deprecated in \pkg{FFTrees}:
\code{fft()} (use \code{\link{FFTrees}()})
}
\keyword{internal}

