% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_data.R
\name{describe_data}
\alias{describe_data}
\title{Describe data}
\usage{
describe_data(data, data_name, criterion_name, baseline_value)
}
\arguments{
\item{data}{A data frame.}

\item{data_name}{A character string specifying the name of the data.}

\item{criterion_name}{A character string specifying the criterion name.}

\item{baseline_value}{The value used as a baseline.}
}
\value{
A data frame with the descriptive statistics.
}
\description{
This function calculates various descriptive statistics for a given dataset.
}
\examples{
data(heartdisease)
describe_data(heartdisease, "heartdisease", "diagnosis", TRUE)
}
