% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_gfft.R
\name{drop_nodes}
\alias{drop_nodes}
\title{Drop a node from an FFT definition}
\usage{
drop_nodes(fft, nodes = NA, quiet = FALSE)
}
\arguments{
\item{fft}{One FFT definition
(as a data frame in tidy format, with one row per node).}

\item{nodes}{The FFT nodes to drop (as an integer vector).
Default: \code{nodes = NA}.}

\item{quiet}{Hide feedback messages (as logical)?
Default: \code{quiet = FALSE}.}
}
\value{
One FFT definition
(as a data frame in tidy format, with one row per node).
}
\description{
\code{drop_nodes} drops or removes
a node from an existing FFT (by deleting the
corresponding row in its definition in
the tidy data frame format).

When dropping the final node,
the last remaining node gets a second exit.

Duplicates in \code{nodes} are dropped only once
(i.e., not incrementally) and \code{nodes} not in
the range \code{1:nrow(fft)} are ignored.
Dropping all nodes yields an error.
}
\seealso{
\code{\link{get_fft_df}} for getting the FFT definitions of an \code{FFTrees} object;
\code{\link{read_fft_df}} for reading one FFT definition from tree definitions;
\code{\link{add_fft_df}} for adding FFTs to tree definitions;
\code{\link{FFTrees}} for creating FFTs from and applying them to data.

Other tree definition and manipulation functions: 
\code{\link{add_fft_df}()},
\code{\link{get_fft_df}()},
\code{\link{read_fft_df}()},
\code{\link{write_fft_df}()}
}
\concept{tree definition and manipulation functions}
\concept{tree trimming functions}
