% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_abc.R
\name{fact_clean}
\alias{fact_clean}
\title{Clean factor variables in prediction data}
\usage{
fact_clean(data.train, data.test, show.warning = T)
}
\arguments{
\item{data.train}{A training dataset}

\item{data.test}{A testing dataset}

\item{show.warning}{logical}
}
\description{
Clean factor variables in prediction data
}
