% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factclean_function.R
\name{factclean}
\alias{factclean}
\title{Calculates predictions from CART}
\usage{
factclean(data.train, data.test, show.warning = T)
}
\arguments{
\item{data.train}{A training dataset}

\item{data.test}{A testing dataset}

\item{formula}{a formula}
}
\description{
Calculates predictions from CART
}

