% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertility_doc.R
\docType{data}
\name{fertility}
\alias{fertility}
\title{Fertility data set.}
\format{
A data frame containing 100 rows and 10 columns.
}
\source{
https://archive.ics.uci.edu/ml/datasets/Fertility
}
\usage{
fertility
}
\description{
Fertility data set.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
