% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Fftm}
\alias{Fftm}
\alias{Fftm.default}
\title{Wrapper function fftm}
\usage{
Fftm(formula, ...)

\method{Fftm}{default}(formula, method, data, criterion = NULL,
  autoprune = TRUE, xvalidate.parts = 2, sdtfun = NULL, cuelist = NULL,
  ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{...}{Additional parameter which should be passed to called \code{method}.}

\item{method}{stringvector which specifies the method. \link{bruteforcemaximize}, \link{maximize}, \link{montecarlo} or \link{chase}}

\item{data}{Dataframe containing the variables in the model.}

\item{criterion}{\code{\link{cue}}; OPTIONAL: If criterion in \code{data} isn't binary, use a predefined criterion cue.}

\item{autoprune}{logical: If set to true \code{\link{prune}} is called before returning the tree or all trees of the list. Default: TRUE}

\item{xvalidate.parts}{integer, how many party of \code{data} should be generated (odd-even method) to crossvalidate the model}

\item{sdtfun}{optional function with 4 parameters (\code{hits}, \code{falsealarms}, \code{misses}, \code{correctrejections}). Has to return a named vector. The returned value can be maximized.}

\item{cuelist}{optional list with cues. The \code{name}s of the cues have to match the columns provided in \code{data}.}
}
\value{
ANY; Depending on used method, return value can either be a \code{\link{list}}, or a \code{\link{fftree}}.
}
\description{
Wrapper function fftm

\code{Fftm} is used to fit Fast- and Frugal tree models, specified by giving a
symbolic description of the predictors and the criterion.
}
\seealso{
\code{\link{Fftm.default}}

\code{\link{bruteforcemaximize}}, \code{\link{maximize}}, \code{\link{chase}}, \code{\link{montecarlo}}
}

