% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fftm.Titanic.data}
\alias{fftm.Titanic.data}
\alias{fftm.UCBAdmissions.data}
\title{Prepares Titanic data for fftm}
\usage{
fftm.Titanic.data()

fftm.UCBAdmissions.data()
}
\value{
data.frame with modified titanic data

data.frame with modified titanic data
}
\description{
Prepares Titanic data for fftm

Prepares Titanic data for fftm
}
\seealso{
\code{\link{Fftm}}
...

\code{\link{Fftm}}
...
}

