% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{fftree-class}
\alias{Fftree}
\alias{fftree}
\alias{fftree-class}
\title{Class fftree}
\usage{
Fftree(criterion, fftcues, treename = "")
}
\arguments{
\item{criterion}{cue}

\item{fftcues}{list with cues. The order of the list represents the hierarchy of the \code{cues}}

\item{treename}{charactervector, name of the tree (just for description, without function).}
}
\value{
fftree S4 object
}
\description{
Class \code{fftree} is a tree, consisting of \code{cues} and a \code{criterion}.

A \code{fftree} is an S4 object, that represents a Fast- and Frugal tree.
Fast- and Frugal trees use \code{cues} to predict data. It's success is measured
at a \code{criterion}, which is also defined as a cue.
}
\seealso{
\code{\link{cue}}
}

