% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_abc.R
\docType{package}
\name{FFTrees-package}
\alias{FFTrees-package}
\title{FFTrees: Generate, Visualise, and Evaluate Fast-and-Frugal Decision Trees}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create, visualize, and test fast-and-frugal decision trees (FFTs) using the algorithms and methods described by Phillips, Neth, Woike & Gaissmaier (2017), \doi{10.1017/S1930297500006239}. FFTs are simple and transparent decision trees for solving binary classification problems. FFTs can be preferable to more complex algorithms because they require very little information, are easy to understand and communicate, and are robust against overfitting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=FFTrees}
  \item \url{https://github.com/ndphillips/FFTrees/}
  \item Report bugs at \url{https://github.com/ndphillips/FFTrees/issues}
}

}
\author{
\strong{Maintainer}: Hansjoerg Neth \email{h.neth@uni.kn} (\href{https://orcid.org/0000-0001-5427-3141}{ORCID})

Authors:
\itemize{
  \item Nathaniel Phillips \email{Nathaniel.D.Phillips.is@gmail.com} (\href{https://orcid.org/0000-0002-8969-7013}{ORCID})
  \item Jan Woike (\href{https://orcid.org/0000-0002-6816-121X}{ORCID})
  \item Wolfgang Gaissmaier (\href{https://orcid.org/0000-0001-6273-178X}{ORCID})
}

}
\keyword{internal}
