% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesguide_function.R
\name{FFTrees.guide}
\alias{FFTrees.guide}
\title{Opens the FFTrees package guide}
\usage{
FFTrees.guide()
}
\value{
No return value, called for side effects
}
\description{
Opens the FFTrees package guide
}
