% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_ffttowords.R
\name{fftrees_ffttowords}
\alias{fftrees_ffttowords}
\title{Describes a fast-and-frugal tree (FFT) in words.}
\usage{
fftrees_ffttowords(x = NULL, digits = 2)
}
\arguments{
\item{x}{An \code{FFTrees} object created with \code{\link{FFTrees}}.}

\item{digits}{How many digits to round numeric values (as integer)?}
}
\value{
A list of string vectors.
}
\description{
Describes a fast-and-frugal tree (FFT) in words.
}
\examples{

heart.fft <- FFTrees(diagnosis ~ .,
  data = heartdisease,
  decision.labels = c("Healthy", "Disease")
)

inwords(heart.fft)

}
