% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_grow_fan.R
\name{fftrees_grow_fan}
\alias{fftrees_grow_fan}
\title{Grows fast-and-frugal trees (FFTs) using the \code{fan} algorithm.}
\usage{
fftrees_grow_fan(x, repeat.cues = TRUE)
}
\arguments{
\item{x}{An \code{FFTrees} object.}

\item{repeat.cues}{logical.}
}
\description{
Grows fast-and-frugal trees (FFTs) using the \code{fan} algorithm.
}
