% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_ranktrrees.R
\name{fftrees_ranktrees}
\alias{fftrees_ranktrees}
\title{Rank trees by goal}
\usage{
fftrees_ranktrees(x, data = "train")
}
\arguments{
\item{x}{FFTrees.}

\item{data}{character.}
}
\description{
Rank trees by goal
}
