% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_wordstofftrees.R
\name{fftrees_wordstofftrees}
\alias{fftrees_wordstofftrees}
\title{Converts text describing an FFT into an FFT definition.}
\usage{
fftrees_wordstofftrees(x, my.tree)
}
\arguments{
\item{x}{An \code{FFTrees} object.}

\item{my.tree}{string. A verbal string defining an FFT.}
}
\value{
An \code{FFTrees} object with a new definition defined by \code{my.tree}.
}
\description{
Converts text describing an FFT into an FFT definition.
}
