% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_abc.R
\name{get_fft_df}
\alias{get_fft_df}
\title{Get FFT definitions (from an \code{FFTrees} object \code{x})}
\usage{
get_fft_df(x)
}
\arguments{
\item{x}{An \code{FFTrees} object.}
}
\value{
A set of FFT definitions (as a \code{data.frame}/\code{tibble}).
}
\description{
\code{get_fft_df} gets the FFT definitions
of an \code{FFTrees} object \code{x}
(as a \code{data.frame}).
}
\details{
In addition to looking up \code{x$trees$definitions},
\code{get_fft_df} verifies that the FFT definitions
are valid (given current settings).
}
\seealso{
\code{\link{FFTrees}} for creating FFTs from and applying them to data.

Other utility functions: 
\code{\link{get_best_tree}()},
\code{\link{get_exit_type}()}
}
\concept{utility functions}
