% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCueEfficiency}
\alias{getCueEfficiency}
\title{Converts \code{fftree} to a \code{data.frame} with it's efficiency values}
\usage{
getCueEfficiency(tree, hideRest = FALSE, includeRest = FALSE)
}
\arguments{
\item{tree}{fftree}

\item{hideRest}{logical: Should "rest" be hidden? Default FALSE.}

\item{includeRest}{logical: Should "rest" be added to last cue (because last cue automatically predicts rest)}
}
\value{
data.frame
}
\description{
Converts \code{fftree} to a \code{data.frame} with it's efficiency values
}
\details{
All values are cue-related. \code{fPot} is the number of unpredicted criterions, \code{cPot} is the quantity of unpredicted criterions WITHIN the spectrum of the cue, \code{boldness} represents the cPot/fPot ratio, \code{cEff} the correct predictions/cPot ratio. All columns can be used by \code{\link{chase}}.
}
\seealso{
\code{\link{Sdt},\link{fftree}}
}

