% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSingleCueEfficiency}
\alias{getSingleCueEfficiency}
\title{Returns a vector with the cue efficiency values of a specific cue}
\usage{
getSingleCueEfficiency(tree, id = length(tree@fftcues), includeRest = TRUE)
}
\arguments{
\item{tree}{fftree}

\item{id}{integer. Specifies which \code{cue} within \code{fftree} is meant. Default: Last cue.}

\item{includeRest}{logical: Should "rest" be added to last cue (because last cue automatically predicts rest)? Default: TRUE}
}
\value{
Vector with "cEff", "triv.", "lazy", "prunable"
}
\description{
Returns a vector with the cue efficiency values of a specific cue
}
\details{
All values are cue-related.
}
\seealso{
\code{\link{Sdt},\link{fftree}, \link{getCueEfficiency}}
}

