% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTreeDevelopment}
\alias{getTreeDevelopment}
\title{Converts \code{fftree} to a \code{data.frame}}
\usage{
getTreeDevelopment(tree)
}
\arguments{
\item{tree}{fftree}
}
\value{
data.frame
}
\description{
Converts \code{fftree} to a \code{data.frame}
}
\seealso{
\code{\link{Sdt}}, \code{\link{fftree}}
}

