% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growFFTrees_function.R
\name{grow.FFTrees}
\alias{grow.FFTrees}
\title{Grows fast-and-frugal trees using an algorithm specified by \code{algorithm}.}
\usage{
grow.FFTrees(x)
}
\arguments{
\item{x}{FFTrees. An FFTrees object}
}
\description{
Grows fast-and-frugal trees using an algorithm specified by \code{algorithm}.
}
