% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartfff_doc.R
\docType{data}
\name{heart.fff}
\alias{heart.fff}
\title{An FFForest object trained on the heartdisease dataset.}
\format{An FFForest object}
\usage{
heart.fff
}
\description{
An FFForest object trained on the heartdisease dataset.
}
\details{
The code used to generate this object was \code{heart.fff <- FFForest(diagnosis ~., data = heartdisease, ntree = 100, train.p = .5, cpus = 4)}
}
\keyword{datasets}
