% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{heart.test}
\alias{heart.test}
\title{Heart disease testing data}
\format{
A data frame containing 153 rows and 14 columns (see \code{\link{heartdisease}} for details).
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Heart+Disease}
}
\usage{
heart.test
}
\description{
Testing data for a \code{\link{heartdisease}} data.
This subset is used to test the prediction performance of a model trained on the \code{\link{heart.train}} data.
The dataset \code{\link{heartdisease}} contains both datasets.
}
\seealso{
\code{\link{heartdisease}} dataset.

Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
