% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{heart.train}
\alias{heart.train}
\title{Heart disease training data}
\format{
A data frame containing 150 rows and 14 columns (see \code{\link{heartdisease}} for details).
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Heart+Disease}
}
\usage{
heart.train
}
\description{
Training data for a binary prediction model (here: FFT) on (a subset of) the \code{\link{heartdisease}} data.
The complementary subset for model testing is \code{\link{heart.test}}.
The data in \code{\link{heartdisease}} contains both subsets.
}
\seealso{
\code{\link{heartdisease}} dataset.

Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
