% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartdisease_doc.R
\docType{data}
\name{heartdisease}
\alias{heartdisease}
\title{heartdisease}
\format{A data frame containing 303 rows and 14 columns
\describe{
  \item{age}{Numeric}
  \item{sex}{Factor}
  \item{cp}{Factor}
  \item{trestbps}{Numeric}
  \item{chol}{Numeric}
  \item{fbs}{Factor}
  \item{restecg}{Factor}
  \item{thalach}{Numeric}
  \item{exang}{Factor}
  \item{oldpeak}{Numeric}
  \item{slope}{Numeric}
  \item{ca}{Numeric}
  \item{thal}{Factor}
  \item{diagnosis}{Factor. 1 = Heart disease, 0 = No Heart disease}



  ...
}}
\source{
http://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Diagnostic)
}
\usage{
heartdisease
}
\description{
Data for 303 patients tested for heart disease
}
\keyword{datasets}

