% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_algorithm.R
\name{heuristic.algorithm}
\alias{heuristic.algorithm}
\title{Heuristic algorithms for building FFTs originally developed by Martignon, Katsikopoulos & Woike (2008)}
\usage{
heuristic.algorithm(formula, data, max.levels = NULL, algorithm = "max")
}
\arguments{
\item{formula}{formula. A formula specifying a logical criterion as a function of 1 or more predictors.}

\item{data}{dataframe. A training dataset.}

\item{max.levels}{integer. The maximum number of levels considered for the trees.}

\item{algorithm}{string. Either 'max' or 'zigzag'}
}
\value{
A definition of an FFT
}
\description{
Heuristic algorithms for building FFTs originally developed by Martignon, Katsikopoulos & Woike (2008)
}
