% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_doc.R
\docType{data}
\name{income}
\alias{income}
\title{Income dataset}
\format{A data frame containing 32561 rows and 15 columns
\describe{
  \item{age}{Numeric}
  \item{workclass}{Factor}
  \item{fnlwgt}{Numeric}
  \item{education}{Factor}
  \item{education.num}{Numeric}
  \item{marital.status}{Factor}
  \item{occupation}{Factor}
  \item{relationship}{Factor}
  \item{race}{Factor}
  \item{sex}{Factor}
  \item{capital.gain}{Numeric}
  \item{capital.loss}{Numeric}
  \item{hours.per.week}{Numeric}
  \item{native.country}{Factor}
  \item{income.gt50}{Factor - 1 = >= 50K, 0 = < 50K}



  ...
}}
\source{
http://archive.ics.uci.edu/ml/datasets/Census+Income
}
\usage{
income
}
\description{
Census Data used to predict who has an income greater than 50K
}
\keyword{datasets}
