% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irisv_doc.R
\docType{data}
\name{iris.v}
\alias{iris.v}
\title{Iris data set.}
\format{
A data frame containing 150 rows and 4 columns.
}
\source{
https://archive.ics.uci.edu/ml/datasets/Iris
}
\usage{
iris.v
}
\description{
Iris data set.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
