% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isCueLazy}
\alias{isCueLazy}
\title{Checks if a specific \code{cue} in \code{fftree} is lazy}
\usage{
isCueLazy(tree, cl = tree@fftcues, n = length(cl))
}
\arguments{
\item{tree}{\code{\link{fftree}}}

\item{cl}{Optional list with cues. Default is the cue list of the given tree}

\item{n}{Optional, which cue of tree should be analysed? Default is last cue}
}
\value{
logical; TRUE if last cue is lazy, else FALSE
}
\description{
A cue is defined as \code{lazy}, if it's \code{cPot} is zero. It can be \code{pruned}
without changing tree- outcomes.
}
\seealso{
\code{\link{getCueEfficiency}}, \code{\link{isCueTrivial}}, \code{\link{isCuePrunable}}
}

