% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isCuePrunable}
\alias{isCuePrunable}
\title{Checks if a specific \code{cue} in \code{fftree} is prunable}
\usage{
isCuePrunable(tree, asMatrix = FALSE, n = 1:length(tree@fftcues))
}
\arguments{
\item{tree}{\code{\link{fftree}}}

\item{asMatrix}{optional logical. If set to TRUE a data.frame is returned, with all information about possible prunings}

\item{n}{vector or single value}
}
\value{
logical vector/data.frame
}
\description{
A cue is defined as \code{prunable}, if it's \code{lazy} or it's follower is \code{trivial}
}
\seealso{
\code{\link{getCueEfficiency}}, \code{\link{isCueTrivial}}, \code{\link{isCueLazy}}
}

