% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isCueTrivial}
\alias{isCueTrivial}
\title{Checks if a specific \code{cue} in \code{fftree} is trivial}
\usage{
isCueTrivial(tree, cl = tree@fftcues, n = length(cl))
}
\arguments{
\item{tree}{\code{\link{fftree}}}

\item{cl}{Optional list with cues. Default is the cue list of the given tree}

\item{n}{Optional, which cue of tree should be analysed? Default is last cue}
}
\value{
logical; TRUE if last cue is trivial, else FALSE
}
\description{
The last cue is defined as trivial, if the last cue could replace the
second to last, without changing tree- outcomes.
}
\seealso{
\code{\link{isCueLazy}}, \code{\link{isCuePrunable}}
}

