% Generated by roxygen2 (4.0.2): do not edit by hand
\name{montecarlo}
\alias{fftm.montecarlo}
\alias{montecarlo}
\title{Maximizes a \code{\link{Sdt}} value by \code{monte carlo} method}
\usage{
fftm.montecarlo(tree, whatToMaximize, blocksize, threshold, cluster = NULL,
  cores = 1)
}
\arguments{
\item{tree}{\code{\link{fftree}}}

\item{whatToMaximize}{stringvector; Which parameters of the tree should be chased? The index of the vector represents the sequence.  See \link{Sdt} for possible values.}

\item{blocksize}{integer; How many random trees should be generated, before \code{threshold} is checked?}

\item{threshold}{double; How much of an improvement in \code{whatToMaximize} should be found to generate another block?}

\item{cluster}{If you have a HPC cluster, specifiy configuration here. See \code{\link{makeCluster}} from \code{snow} package.}

\item{cores}{Integer; If multithreading is wanted (recommended for bigger computations!), you can specify your cores here.}
}
\value{
\code{\link{Fftree}}
}
\description{
Takes all \code{\link{cue}s} in \code{\link{fftree}} and builds \code{blocksize} random generated fftrees.
If no better fftree is found, or the the value didn't improve greater than \code{threshold}, the algorithm finishes.
Use \code{\link{Fftm}} instead of calling this method directly.
}
\details{
You can chase all values provided by \code{\link{Sdt}}
}
\examples{
rawdata <- fftm.Titanic.data()
tree <- Fftm(Survived ~ Age + Sex + Class, "montecarlo",rawdata, whatToMaximize = "percCorr", blocksize = 1000, threshold=.001, cores=4)
}
\seealso{
\code{\link{Fftm}}, \code{\link{bruteforcemaximize}}, \code{\link{chase}}, \code{\link{maximize}}
}

