% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkplot.R
\name{network_plot}
\alias{network_plot}
\title{Creates a network plot. Code taken from Dirk Wulff (www.dirkwulff.org)}
\usage{
network_plot(edges, nodesize = 0.1, edgesize = 1, mincon = 0)
}
\arguments{
\item{edges}{dataframe. A dataframe showing}

\item{nodesize}{numeric. Nodesize adjustment}

\item{edgesize}{numeric. Edgesize adjustment}

\item{mincon}{integer. Minimum connection cutoff}
}
\description{
Creates a network plot. Code taken from Dirk Wulff (www.dirkwulff.org)
}

