% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFFForest.R
\name{plot.FFForest}
\alias{plot.FFForest}
\title{Creates a network plot. Code taken from Dirk Wulff (www.dirkwulff.org)}
\usage{
\method{plot}{FFForest}(x, nodesize = 0.1, edgesize = 1, mincon = 0, ...)
}
\arguments{
\item{x}{FFForest. An FFForest object created from FFForest()}

\item{nodesize}{numeric. Nodesize adjustment}

\item{edgesize}{numeric. Edgesize adjustment}

\item{mincon}{integer. Minimum connection cutoff}

\item{...}{currently ignored}
}
\description{
Creates a network plot. Code taken from Dirk Wulff (www.dirkwulff.org)
}

