% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFFTrees_function.R
\name{plot.FFTrees}
\alias{plot.FFTrees}
\title{Plots an FFTrees object.}
\usage{
\method{plot}{FFTrees}(
  x = NULL,
  data = "train",
  what = "tree",
  tree = 1,
  main = NULL,
  hlines = TRUE,
  cue.labels = NULL,
  decision.labels = NULL,
  cue.cex = NULL,
  threshold.cex = NULL,
  decision.cex = 1,
  comp = TRUE,
  stats = TRUE,
  show.header = NULL,
  show.tree = NULL,
  show.confusion = NULL,
  show.levels = NULL,
  show.roc = NULL,
  show.icons = NULL,
  show.iconguide = NULL,
  label.tree = NULL,
  label.performance = NULL,
  n.per.icon = NULL,
  which.tree = NULL,
  level.type = "bar",
  decision.names = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{FFTrees} object created by the \code{\link{FFTrees}} function.}

\item{data}{The data to be plotted (as a string), either \code{"train"} or \code{"test"}. The corresponding dataset in object \code{x} will be used.}

\item{what}{What should be plotted (as a string)?
\code{'tree'} (the default) shows one tree (specified by \code{'tree'}).
\code{'cues'} shows the marginal accuracy of cues in ROC space,
\code{"roc"} shows the performance of tree(s) (and comparison algorithms) in ROC space.}

\item{tree}{The tree to be plotted (as an integer, only valid when the corresponding tree argument is non-empty).
To plot the best training (or test) tree with respect to the \code{goal} specified during FFT construction,
use \code{"best.train"} or \code{"best.test"}.}

\item{main}{The main plot label (as character string).}

\item{hlines}{Should horizontal panel separation lines be shown (as logical)?}

\item{cue.labels}{An optional string of labels for the cues / nodes (as character vector).}

\item{decision.labels}{A character vector of length 2 indicating the content-specific names for noise and signal predictions/exits.}

\item{cue.cex}{The size of the cue labels (as numeric).}

\item{threshold.cex}{The size of the threshold labels (as numeric).}

\item{decision.cex}{The size of the decision labels (as numeric).}

\item{comp}{Should the performance of competitive algorithms (e.g.; logistic regression, random forests etc.) be shown in the ROC plot (if available, as logical)?}

\item{stats}{Should statistical information be plotted (as logical)?
If \code{FALSE}, only the tree (without any reference to statistics) will be plotted.}

\item{show.header}{Show header with basic data properties (in top panel, as logical)?}

\item{show.tree}{Show nodes and exits of FFT (in middle panel, as logical)?}

\item{show.confusion}{Show 2x2 confusion matrix (in bottom panel, as logical)?}

\item{show.levels}{Show performance levels (in bottom panel, as logical)?}

\item{show.roc}{Show ROC curve (in bottom panel, as logical)?}

\item{show.icons}{Show exit cases as icon arrays (in middle panel, as logical)?}

\item{show.iconguide}{Show icon guide (in middle panel, as logical)?}

\item{label.tree}{Label for the FFT (optional, as character string).}

\item{label.performance}{Labels for the performance section (optional, as character string).}

\item{n.per.icon}{Number of cases per icon (as numeric).}

\item{which.tree}{deprecated argument, included for backwards compatibility, use \code{"tree"} instead.}

\item{level.type}{How should bottom levels be drawn (as a string)? Can be \code{"bar"} (the default) or \code{"line"}.}

\item{decision.names}{deprecated argument.}

\item{...}{Currently ignored.}
}
\value{
A plot visualizing and describing an FFT.
}
\description{
\code{plot.FFTrees} plots an \code{FFTrees} object created by the \code{\link{FFTrees}} function.

\code{plot.FFTrees} is the main plotting function of the \strong{FFTrees} package and
called when evaluating the generic \code{\link{plot}} on an \code{FFTrees} object.

\code{plot.FFTrees} visualizes a selected FFT, key data characteristics, and metrics of classification performance.

Many aspects of the plot (e.g., its panels) and the FFT's appearance (e.g., labels of its nodes and exits) can be customized by setting corresponding arguments.
}
\examples{
# Create FFTs (for heartdisease data):
heart.fft <- FFTrees(formula = diagnosis ~ .,
                     data = heartdisease
                     )

# Visualize the default FFT (Tree #1):
plot(heart.fft,
     main = "Heart Disease Diagnosis",
     decision.labels = c("Absent", "Present")
     )

# Visualize FFT #2 (with customized labels):
plot(heart.fft,
     tree = 2,
     main = "An FFT for heart disease diagnosis",
     cue.labels = c("1. thal?", "2. cp?", "3. ca?", "4. exang"),
     decision.labels = c("ok", "sick"),
     show.header = FALSE,
     show.confusion = FALSE,
     show.levels = FALSE,
     show.roc = FALSE
     )

# Visualize cue accuracies:
plot(heart.fft, what = "cues")

# For more details, see
vignette("FFTrees_plot", package = "FFTrees")

}
\seealso{
\code{\link{print.FFTrees}} for printing summary information of FFTs;
\code{\link{showcues}} for plotting cue accuracies;
\code{\link{FFTrees}} for creating FFTs from and applying them to data.

Other plot functions: 
\code{\link{showcues}()}
}
\concept{plot functions}
