% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot}
\alias{plot}
\alias{plot.fftree}
\title{Plots \code{fftree}}
\usage{
\method{plot}{fftree}(object, showBoldness = FALSE, ...)
}
\arguments{
\item{object}{fftree}

\item{showBoldness}{logical, set to TRUE, if the boldness on the x-axis should be shown}

\item{...}{unused. Further parameter.}
}
\description{
Plots \code{fftree}
}
\seealso{
\code{\link{fftree}}
}

