% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predict,fftree-method}
\alias{predict,fftree-method}
\title{Predicts data with \code{fftree}}
\usage{
\S4method{predict}{fftree}(object, newdata = NULL, type = "response", ...)
}
\arguments{
\item{object}{\code{\link{fftree}}}

\item{newdata}{OPTIONAL data.frame. Columns of \code{newdata} have to match the \code{names} of the \code{\link{cue}}s in the \code{\link{fftree}}. If \code{newdata} is \code{NULL}, the fitting values of the \code{fftree} are shown.}

\item{type}{OPTIONAL string. Has to be "response" or "class". Default is "response".}
}
\description{
Predicts data with \code{fftree}
}
\seealso{
\code{\link{fftree}}
}

