% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFFForest_function.R
\name{predict.FFForest}
\alias{predict.FFForest}
\title{Predict data from an FFForest object}
\usage{
\method{predict}{FFForest}(object = NULL, data = NULL, threshold = 0.5,
  ...)
}
\arguments{
\item{object}{FFForest. An FFForest object created from the FFForest() function.}

\item{data}{dataframe. A dataframe of test data}

\item{threshold}{numeric. A threshold value}

\item{...}{Additional arguments passed on to \code{predict()}}
}
\value{
A logical vector of predictions
}
\description{
Predict data from an FFForest object
}

