% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predict}
\alias{predict}
\title{Predicts value with parameter of cue}
\usage{
\S4method{predict}{cue}(object, newdata = NULL)
}
\arguments{
\item{object}{cue}

\item{newdata}{numerical vector with data which should be predictes}
}
\value{
logical. TRUE/FALSE
}
\description{
If parameters of cue have been changed manually, this has to be called. Internal.
}
\seealso{
\code{\link{cue}} \code{\link{cue-class}}
}

