% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFFForest_function.R
\name{print.FFForest}
\alias{print.FFForest}
\title{Prints summary information from an FFForest x}
\usage{
\method{print}{FFForest}(x = NULL, ...)
}
\arguments{
\item{x}{FFForest. An FFForest x created from FFForest()}

\item{...}{additional arguments passed to print.}
}
\description{
Printing function for an FFTrees x
}
