% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfft_function.R
\name{print.fft}
\alias{print.fft}
\title{Draws (and creates) a FFT.}
\usage{
\method{print}{fft}(x = NULL, ...)
}
\arguments{
\item{x}{A fft object created from fft()}

\item{...}{additional arguments passed to print.}
}
\description{
Printing function for an fft object
}

