% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFFTrees_function.R
\name{print.FFTrees}
\alias{print.FFTrees}
\title{Prints summary information of fast-and-frugal trees (FFTs).}
\usage{
\method{print}{FFTrees}(x = NULL, tree = 1, ...)
}
\arguments{
\item{x}{An \coce{FFTrees} object created by \code{\link{FFTrees}}.}

\item{tree}{The tree to explore (as integer).}

\item{...}{additional arguments passed to \code{print}.}
}
\value{
Prints summary information about an FFT to the console.
}
\description{
\code{print.FFTrees} provides summary information
on FFTs from an \code{FFTrees} object.
}
\seealso{
\code{\link{plot.FFTrees}} for plotting FFTs;
\code{\link{inwords}} for obtaining a verbal description of FFTs;
\code{\link{FFTrees}} for creating FFTs from data.
}
