% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prune}
\alias{prune}
\alias{prune.fftree}
\title{Prunes a \code{cue} from given \code{fftree}}
\usage{
\method{prune}{fftree}(tree, n = NULL)
}
\arguments{
\item{tree}{fftree.}

\item{n}{integer/integer vector. Index of \code{cue}, which should be pruned, If set to "NULL", all prunable cues will be pruned automaticly. See \code{\link{isCuePrunable}} for details.}
}
\value{
fftree.
}
\description{
Prunes a \code{cue} from given \code{fftree}
}
\seealso{
\code{\link{cue}},\code{\link{fftree}}, \code{\link{isCuePrunable}}, \code{\link{isCueLazy}}, \code{\link{isCueTrivial}}
}

