% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Sdt.fftree}
\alias{Sdt.fftree}
\title{Converts \code{fftree} to a \code{SDT} vector}
\usage{
\method{Sdt}{fftree}(tree)
}
\arguments{
\item{tree}{fftree}
}
\value{
sdt vector
}
\description{
The Sdt- function needs \code{hits}, \code{false alarms}, \code{correct rejections} and \code{misses}.
By comparing the \code{prediciton} aka the \code{result} of the tree with the criterion,
these values can calculated.
}
\seealso{
\code{\link{Sdt}}
}

