% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Sdt}
\alias{Sdt}
\alias{Sdt.default}
\alias{Sdt.logical}
\title{Wrapper function sdt}
\usage{
Sdt(hi, ...)

\method{Sdt}{default}(hi, fa, mi, cr)

\method{Sdt}{logical}(criterion, prediction)
}
\arguments{
\item{hi}{numeric; hits / true positives}

\item{...}{further parameter}

\item{fa}{numeric; false alarms / false positives}

\item{mi}{numeric; misses / false negatives}

\item{cr}{numeric; correct rejection / true negatives}

\item{criterion}{logical vector}

\item{prediction}{logical vector}
}
\value{
numeric vector with signal-detection values

numeric vector with signal-detection values
}
\description{
Wrapper function sdt

Creates a 'Signal Detection Theory' vector

Creates a 'Signal Detection Theory' vector
}
\details{
This function returns: hitrate (sensitivity/TPR), specifity (true negative rate/SPC), false alarm rate (fall-out/FPR), false discovery rate (FDR),
an estimated d' (qnorm(hitrate)-qnorm(false alarm rate)) and the MCC, the
"Matthews correlation efficient".

Some results are adjusted, to make them calculatable. If one of the contingency-values \code{hi}, \code{fa}, \code{mi} or \code{cr} equals zero,
all of them will gain .25: \code{Sdt(1, 0, 2, 4)} equals \code{Sdt(1.25, .25, 2.25, 4.25)}.
The denominator of the Matthews correlation coefficient is adjusted to 1 if \code{(hi + fa) == 0}, \code{(hi + mi) == 0}, \code{(fa + cr) == 0} or \code{(cr + mi) == 0}.
}
\references{
\url{http://kangleelab.com/signal detection theory.html} \url{http://en.wikipedia.org/wiki/Matthews_correlation_coefficient}
}
\seealso{
\code{\link{Sdt.fftree}}

\code{\link{Sdt.fftree}}
}

