% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show}
\alias{show}
\alias{toString}
\title{Shows \code{cue}}
\usage{
\S4method{show}{cue}(object)

\S4method{show}{fftree}(object)
}
\description{
Shows \code{cue}

Shows \code{fftree}
}
\seealso{
\code{\link{cue}}
...

\code{\link{Sdt},\link{fftree}, \link{getCueEfficiency}, \link{getTreeDevelopment}}
}

