% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfftplot_function.R
\name{simfftplot}
\alias{simfftplot}
\title{Plots the results of the simfft function}
\usage{
simfftplot(simfft.result, roc = F, lr = F, cart = F,
  which.data = "test")
}
\arguments{
\item{simfft.result}{A dataframe returned by the simfft function}

\item{roc}{A logical value indicating whether to plot an pseudo ROC curve (roc = T) or a plot showing how often cues were used (roc = F)}

\item{lr, cart}{Logical values indicating whether or not to plot the logistic regression and CART ROCs (ignored if roc = F)}

\item{which.data}{Which data should be plotted in the roc? Can be "train" or "test" (this is ignored if roc = F)}
}
\description{
Plots the results of the simfft function
}

