% Generated by roxygen2 (4.0.2): do not edit by hand
\name{splitvalanalysis}
\alias{splitvalanalysis}
\title{Returns a SDT table for all splitvalues of a specific independentend variable}
\usage{
splitvalanalysis(criterion, data, name = NULL)
}
\arguments{
\item{criterion}{\code{\link{cue}}, which indicates the criterion to be used}

\item{data}{dataframe or vector. If data is a data.frame, you have to specify the \code{name} of the column provided in the dataframe of \code{data}}

\item{name}{of independent variable in column}
}
\value{
dataframe with SDT values for all possible splitvalues
}
\description{
Returns a SDT table for all splitvalues of a specific independentend variable
}
\seealso{
\code{\link{cue}}, code{\link{Sdt}}
}

