% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryfft_function.R
\name{summary.fft}
\alias{summary.fft}
\title{Returns a summary of an fft object}
\usage{
\method{summary}{fft}(object, ...)
}
\arguments{
\item{object}{An fft object}

\item{...}{Additional arguments}
}
\description{
Returns a summary of an fft object
}

