% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFFTrees_function.R
\name{summary.FFTrees}
\alias{summary.FFTrees}
\title{Summarize an \code{FFTrees} object}
\usage{
\method{summary}{FFTrees}(object, tree = 1, ...)
}
\arguments{
\item{object}{An \code{FFTrees} object.}

\item{tree}{The tree to summarize (as an integer).}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A data frame containing summary information about an FFT.
}
\description{
Summarize an \code{FFTrees} object
}
