% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmpred_function.R
\name{svm.pred}
\alias{svm.pred}
\title{Calculates predictions from support vector machines using the e1071 package}
\usage{
svm.pred(formula, data.train, data.test = NULL, svm.model = NULL)
}
\arguments{
\item{formula}{a formula}

\item{data.train}{dataframe. A training dataset}

\item{data.test}{dataframe. A testing dataset}

\item{svm.model}{An optional existing svm model}
}
\description{
Calculates predictions from support vector machines using the e1071 package
}
\examples{

# Fit svm for the mushrooms dataset

mushrooms.svm <- svm.pred(formula = poisonous ~.,
                          data.train = mushrooms)



}

