% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{toString,cue-method}
\alias{toString,cue-method}
\title{#Converts \code{cue} in named vector.
#
#}
\usage{
\S4method{toString}{cue}(x, suppressNameBrackets = FALSE,
  suppressThen = FALSE, suppressName = FALSE, suppressTest = FALSE)
}
\arguments{
\item{x}{cue}

\item{suppressNameBrackets}{logical}

\item{suppressThen}{logical}

\item{suppressName}{logical}

\item{suppressTest}{logical}

\item{object}{cue
#
#}
}
\value{
numeric vector
#

string
}
\description{
#Converts \code{cue} in named vector.
#
#
}
\seealso{
\code{\link{Cue}} \code{\link{cue-class}}
#...
Converts \code{cue} in readable string format.

\code{\link{Cue}} \code{\link{cue-class}}
}

