% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparent_function.R
\name{transparent}
\alias{transparent}
\title{Makes transparent colors}
\usage{
transparent(orig.col = "red", trans.val = 0.5)
}
\arguments{
\item{orig.col}{a color}

\item{trans.val}{A number between 0 and 1 indicating how transparent to make the color}
}
\description{
Makes transparent colors
}

