% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{update,cue-method}
\alias{update,cue-method}
\title{Refreshes/updated cue- prediction- cache.}
\usage{
\S4method{update}{cue}(object, data)
}
\arguments{
\item{object}{cue}

\item{data}{numerical vector with new prediciton data}
}
\value{
list with S4 cue objects
}
\description{
If parameters of cue have been changed manually, this has to be called. Internal.
}
\seealso{
\code{\link{Cue}} \code{\link{cue-class}}
}

