% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{update}
\alias{update}
\alias{updateFftree}
\title{Updates/Refreshes \code{\link{fftree}}}
\usage{
updateFftree(fftree)

\S4method{update}{fftree}(object, data = NULL)
}
\arguments{
\item{fftree}{fftree}

\item{object}{fftree}

\item{data}{optional dataframe to update the model on}
}
\value{
fftree
}
\description{
Updates/Refreshes \code{\link{fftree}} internal caches and results, like predictions, hit-/false alarm- counters, rest-data etc.
}
\details{
This a somewhat internal function. Use it with caution. Normally one should use the wrapped version \code{update} instead of the direct call \code{updateFftree}.
}
\seealso{
\code{\link{update},\link{fftree}}
...

\code{\link{fftree}}
}

