% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateFftree2}
\alias{updateFftree2}
\title{Updates tree cache}
\usage{
updateFftree2(fftree, cuelist)
}
\arguments{
\item{fftree}{fftree}

\item{cuelist}{list. List with cues.}
}
\value{
fftree
}
\description{
Updates trees internal caches and results without writing new cues into trees internal storage.
}
\details{
Use it with caution. Normally one should use the wrapped version \code{\link{update}}. This specific variant avoids writing the used cues back into the trees internal storage to avoid the list-bottlenecks.
}
\seealso{
\code{\link{update},\link{fftree}}
...
}

