% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHistory.R
\name{updateHistory}
\alias{updateHistory}
\title{Update the history of decisions from trees in an FFTrees object}
\usage{
updateHistory(object = NULL, newdata = NULL, tree.definitions = NULL,
  formula = NULL)
}
\arguments{
\item{object}{FFTrees. An FFTrees object}

\item{newdata}{dataframe. A dataframe of new data that can be used with \code{predict.FFTrees()}}

\item{tree.definitions}{dataframe. Tree definitions (i.e., from an existing FFTrees object)}

\item{formula}{formula. Formula used when building FFTrees.}
}
\description{
Update the history of decisions from trees in an FFTrees object
}
