% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_doc.R
\docType{data}
\name{voting}
\alias{voting}
\title{Voting data}
\format{
A data frame containing 435 rows and 16 columns.
}
\source{
https://archive.ics.uci.edu/ml/datasets/Congressional+Voting+Records
}
\usage{
voting
}
\description{
Voting data
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
