% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine tasting data}
\format{
A data frame containing 6497 rows and 13 columns.
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Wine+Quality}
}
\usage{
wine
}
\description{
Chemical and tasting data from wines in North Portugal.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}}
}
\concept{datasets}
\keyword{datasets}
