% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordstoFFT.R
\name{wordstoFFT}
\alias{wordstoFFT}
\title{Converts text describing an FFT into an FFT definition.}
\usage{
wordstoFFT(input, cue.names, decision.labels = c("false", "true"))
}
\arguments{
\item{input}{string. A string describing an FFT in words (see examples)}

\item{decision.labels}{string. A vector of decision labels}

\item{cues}{string. A vector of cue names}
}
\description{
Converts text describing an FFT into an FFT definition.
}
\examples{

my.tree.def <- wordstoFFT(input = "If age > 55, predict TRUE. If cp equals [a, b, np], predict False. If slope equals [flat, down], predict FALSE. Otherwise, predict False",
                          cue.names = names(heartdisease))


}
