% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Add_Stats}
\alias{Add_Stats}
\title{Adds decision statistics to a dataframe containing hr, cr, mi and fa}
\usage{
Add_Stats(
  data,
  sens.w = 0.5,
  cost.each = NULL,
  cost.outcomes = list(hi = 0, fa = 1, mi = 1, cr = 0)
)
}
\arguments{
\item{data}{dataframe. With named (integer) columns hi, cr, mi, fa}

\item{sens.w}{numeric. Sensitivity weight}

\item{cost.each}{numeric. An optional fixed cost added to all outputs (e.g.; the cost of the cue)}

\item{cost.outcomes}{list. A list of length 4 with names 'hi', 'fa', 'mi', and 'cr' specifying the costs of a hit, false alarm, miss, and correct rejection rspectively. E.g.; \code{cost.outcomes = listc("hi" = 0, "fa" = 10, "mi" = 20, "cr" = 0)} means that a false alarm and miss cost 10 and 20 respectively while correct decisions have no cost.}
}
\description{
Adds decision statistics to a dataframe containing hr, cr, mi and fa
}
