% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applytree_function.R
\name{apply.tree}
\alias{apply.tree}
\title{Applies a fast and frugal tree to a dataset.}
\usage{
apply.tree(data, formula, tree.definitions, sens.w = 0.5)
}
\arguments{
\item{data}{dataframe. A model training dataset. An m x n dataframe containing n cue values for each of the m exemplars.}

\item{formula}{A formula}

\item{tree.definitions}{dataframe. Definitions of one or more trees. The dataframe must contain the columns: cues, classes, thresholds, directions, exits.}

\item{sens.w}{numeric.  A number from 0 to 1 indicating how to weight sensitivity relative to specificity. Only used for calculating wacc values.}
}
\value{
A list of length 4 containing
}
\description{
Applies a fast and frugal tree to a dataset.
}
\examples{


  tree.definitions <- data.frame("tree" = 1,
                                 "cues" = "sex;age",
                                 "thresholds" = "male;adult",
                                 "directions" = "=;="
                                 )


}
