% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_function.R
\name{auc}
\alias{auc}
\title{Calculates AUC (Area under the Curve) using trapezoidal approximation}
\usage{
auc(sens.v, spec.v)
}
\arguments{
\item{sens.v}{a vector of sensitivities}

\item{spec.v}{A vector of specificities}
}
\description{
Calculates AUC (Area under the Curve) using trapezoidal approximation
}
\examples{

# Calculate the AUC for a vector of hit rates and false alarm rates

auc(sens.v = c(.1, .3, .5, .7), spec.v = c(.05, .1, .15, .3))


}
