% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_function.R
\name{auc}
\alias{auc}
\title{Calculates AUC (Area under the Curve)}
\usage{
auc(hr.v, far.v)
}
\arguments{
\item{hr.v}{a vector of hit rates}

\item{far.v}{A vector of false alarm rates}
}
\description{
Calculates AUC (Area under the Curve)
}

