% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_doc.R
\docType{data}
\name{blood}
\alias{blood}
\title{Blood donation dataset}
\format{A data frame containing 748 rows and 5 columns
\describe{
  \item{recency}{Numeric}
  \item{frequency}{Numeric}
  \item{total}{Numeric}
  \item{time}{Numeric}
  \item{donation.crit}{Numeric}
  ...
}}
\source{
https://archive.ics.uci.edu/ml/datasets/Blood+Transfusion+Service+Center
}
\usage{
blood
}
\description{
Blood donation data
}
\keyword{datasets}

