% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_doc.R
\docType{data}
\name{blood}
\alias{blood}
\title{Blood donation dataset}
\format{
A data frame containing 748 rows and 5 columns
\describe{
  \item{recency}{Months since last donation}
  \item{frequency}{Total number of donations}
  \item{total}{Total blood donated in c.c.}
  \item{time}{Months since first donation}
  \item{donation.crit}{Did he/she donated blood in March 2007?}
  ...
}
}
\source{
https://archive.ics.uci.edu/ml/datasets/Blood+Transfusion+Service+Center
}
\usage{
blood
}
\description{
Blood donation dataset
}
\keyword{datasets}
