% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{breastcancer}
\alias{breastcancer}
\title{Physiological data of patients tested for breast cancer}
\format{
A data frame containing 699 patients (rows) and 9 variables (columns).

\describe{
  \item{thickness}{Clump Thickness}
  \item{cellsize.unif}{Uniformity of Cell Size}
  \item{cellshape.unif}{Uniformity of Cell Shape}
  \item{adhesion}{Marginal Adhesion}
  \item{epithelial}{Single Epithelial Cell Size}
  \item{nuclei.bare}{Bare Nuclei}
  \item{chromatin}{Bland Chromatin}
  \item{nucleoli}{Normal Nucleoli}
  \item{mitoses}{Mitoses}
  \item{diagnosis}{\emph{Criterion}: Absence/presence of breast cancer.

  Values: \code{FALSE} vs. \code{TRUE} (65.0\% vs.\ 35.0\%).}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Original)}

Original creator:

Dr. William H. Wolberg (physician)

University of Wisconsin Hospitals

Madison, Wisconsin, USA
}
\usage{
breastcancer
}
\description{
Physiological data of patients tested for breast cancer
}
\details{
We made the following enhancements to the original data for improved usability:

\itemize{
 \item{The ID number of the cases was excluded.}
 \item{The numeric criterion with value "2" for benign and "4" for malignant was converted to logical TRUE/FALSE.}
 \item{16 cases were excluded because they contained NAs.}
 }

 Other than that, the data remains consistent with the original dataset.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
