% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breastcancer_doc.R
\docType{data}
\name{breastcancer}
\alias{breastcancer}
\title{Dataset: Physiological dataset for 699 patients tested for breast cancer.}
\format{A data frame containing 699 rows and 9 columns
\describe{
  \item{thickness}{Numeric}
  \item{cellsize.unif}{Numeric}
  \item{cellshape.unif}{Numeric}
  \item{adhesion}{Numeric}
  \item{epithelial}{Numeric}
  \item{nuclei.bare}{Numeric}
  \item{chromatin}{Numeric}
  \item{nucleoli}{Numeric}
  \item{mitoses}{Numeric}
  \item{diagnosis}{B = Benign (no cancer), M = Malignant (cancer present))}



  ...
}}
\source{
The diagnostic Wisconsin breast cancer database. http://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Diagnostic)
}
\usage{
breastcancer
}
\description{
Dataset: Physiological dataset for 699 patients tested for breast cancer.
}
\keyword{datasets}

