% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breastcancer_doc.R
\docType{data}
\name{breastcancer}
\alias{breastcancer}
\title{Dataset: Physiological dataset for 699 patients tested for breast cancer.}
\format{A data frame containing 699 rows and 9 columns
\describe{
  \item{thickness}{Clump Thickness}
  \item{cellsize.unif}{Uniformity of Cell Size}
  \item{cellshape.unif}{Uniformity of Cell Shape}
  \item{adhesion}{Marginal Adhesion}
  \item{epithelial}{Single Epithelial Cell Size}
  \item{nuclei.bare}{Bare Nuclei}
  \item{chromatin}{Bland Chromatin}
  \item{nucleoli}{Normal Nucleoli}
  \item{mitoses}{Mitoses}
  \item{diagnosis}{Is cancer present? T or F}



  ...
}}
\source{
https://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Original)
}
\usage{
breastcancer
}
\description{
Dataset: Physiological dataset for 699 patients tested for breast cancer.
}
\keyword{datasets}

