% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{car}
\alias{car}
\title{Car acceptability data}
\format{
A data frame containing 1728 cars (rows) and 7 variables (columns).

\describe{
  \item{buying.price}{Numeric}
  \item{maint.price}{Factor}
  \item{doors}{Factor}
  \item{persons}{Numeric}
  \item{luggage}{Numeric}
  \item{safety}{Factor}
  \item{acceptability}{Factor}

  ...
}
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Car+Evaluation}
}
\usage{
car
}
\description{
A dataset on car evaluations based on basic features, derived from a simple hierarchical decision model.
}
\details{
The (yet to be binarized) criterion variable is a car's \code{acceptability} rating.
}
\references{
Bohanec, M., Rajkovic, V. (1990): Expert system for decision making. Sistemica 1 (1), pp. 145-157.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
