% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_doc.R
\docType{data}
\name{car}
\alias{car}
\title{Car acceptability data}
\format{
A data frame containing 1728 rows and 7 columns
\describe{
  \item{buying.price}{Numeric}
  \item{maint.price}{Factor}
  \item{doors}{Factor}
  \item{persons}{Numeric}
  \item{luggage}{Numeric}
  \item{safety}{Factor}
  \item{acceptability}{Factor}



  ...
}
}
\source{
http://archive.ics.uci.edu/ml/datasets/Car+Evaluation
}
\usage{
car
}
\description{
Car acceptability data
}
\keyword{datasets}
