% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartpred_function.R
\name{cart.pred}
\alias{cart.pred}
\title{Calculates predictions from CART using the rpart package}
\usage{
cart.pred(formula, data.train, data.test = NULL)
}
\arguments{
\item{formula}{a formula}

\item{data.train}{A training dataset}

\item{data.test}{A testing dataset}
}
\description{
Calculates predictions from CART using the rpart package
}

