% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classtable_function.R
\name{classtable}
\alias{classtable}
\title{Calculates classification statistics from binary prediction and truth vectors}
\usage{
classtable(prediction.v, criterion.v)
}
\arguments{
\item{prediction.v}{A binary vector of predictions}

\item{criterion.v}{A binary vector of criterion (true) values}
}
\description{
Calculates classification statistics from binary prediction and truth vectors
}

