% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classtable_function.R
\name{classtable}
\alias{classtable}
\title{Calculates several classification statistics from binary prediction and criterion (e.g.; truth) vectors}
\usage{
classtable(prediction.v, criterion.v)
}
\arguments{
\item{prediction.v}{A binary vector of predictions}

\item{criterion.v}{A binary vector of criterion (true) values}
}
\description{
Calculates several classification statistics from binary prediction and criterion (e.g.; truth) vectors
}

