% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classtable_function.R
\name{classtable}
\alias{classtable}
\title{Calculates several classification statistics from binary prediction and criterion (e.g.; truth) vectors}
\usage{
classtable(prediction.v, criterion.v)
}
\arguments{
\item{prediction.v}{A binary vector of predictions}

\item{criterion.v}{A binary vector of criterion (true) values}
}
\description{
Calculates several classification statistics from binary prediction and criterion (e.g.; truth) vectors
}
\examples{


 # classification statistics for 5 cases
classtable(prediction.v = c(0, 0, 0, 1, 1),
           criterion.v = c(0, 0, 1, 0, 1))

# Same as above, but now is perfect
classtable(prediction.v = c(0, 0, 1, 0, 1),
           criterion.v = c(0, 0, 1, 0, 1))



}

