% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraceptive_doc.R
\docType{data}
\name{contraceptive}
\alias{contraceptive}
\title{Contraceptive use data}
\format{A data frame containing 1473 rows and 10 columns
\describe{
  \item{wife.age}{Numeric}
  \item{wife.edu}{Factor}
  \item{hus.ed}{Factor}
  \item{children}{Numeric}
  \item{wife.rel}{Numeric}
  \item{wife.work}{Factor}
  \item{hus.occ}{Factor}
  \item{sol}{Factor}
  \item{media}{Numeric}
  \item{cont.crit}{numeric}



  ...
}}
\source{
https://archive.ics.uci.edu/ml/datasets/Contraceptive+Method+Choice
}
\usage{
contraceptive
}
\description{
Contraceptive use data
}
\keyword{datasets}

