% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditapproval_doc.R
\docType{data}
\name{creditapproval}
\alias{creditapproval}
\title{Credit approval data}
\format{A data frame containing 690 rows and 15 columns}
\source{
https://archive.ics.uci.edu/ml/datasets/Credit+Approval
}
\usage{
creditapproval
}
\description{
Credit approval data
}
\keyword{datasets}

