% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_data.R
\name{describe_data}
\alias{describe_data}
\title{Describe data}
\usage{
describe_data(data, data_name, criterion_name, baseline_value)
}
\arguments{
\item{data}{A data frame with a criterion variable \code{criterion_name}.}

\item{data_name}{A character string specifying a name for the data.}

\item{criterion_name}{A character string specifying the criterion name.}

\item{baseline_value}{The value in \code{criterion_name} denoting the baseline
(e.g., \code{TRUE} or \code{FALSE}).}
}
\value{
A data frame with the descriptive statistics.
}
\description{
Calculate key descriptive statistics for a given set of data.
}
\examples{
data(heartdisease)
describe_data(heartdisease, "heartdisease",
              criterion_name = "diagnosis",
              baseline_value = TRUE)

}
