% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{factclean}
\alias{factclean}
\title{Does miscellaneous cleaning of prediction datasets}
\usage{
factclean(data.train, data.test, show.warning = T)
}
\arguments{
\item{data.train}{A training dataset}

\item{data.test}{A testing dataset}

\item{show.warning}{...}
}
\description{
Does miscellaneous cleaning of prediction datasets
}
