% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{factclean}
\alias{factclean}
\title{Clean factor variables in prediction data}
\usage{
factclean(data.train, data.test, show.warning = T)
}
\arguments{
\item{data.train}{A training dataset}

\item{data.test}{A testing dataset}

\item{show.warning}{logical}
}
\description{
Clean factor variables in prediction data
}
