% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{fertility}
\alias{fertility}
\title{Fertility data}
\format{
A data frame containing 100 rows and 10 columns.
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Fertility}
}
\usage{
fertility
}
\description{
This dataset describes a sample of 100 volunteers providing a semen sample that was analyzed according to the WHO 2010 criteria.
}
\details{
Sperm concentration are related to socio-demographic data, environmental factors, health status, and life habits.

The binary criterion variable is \code{diagnosis}: Normal (N) vs. altered (O).
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
