% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertility_doc.R
\docType{data}
\name{fertility}
\alias{fertility}
\title{Fertility data set}
\format{A data frame containing 100 rows and 10 columns}
\source{
https://archive.ics.uci.edu/ml/datasets/Fertility
}
\usage{
fertility
}
\description{
Fertility data set
}
\keyword{datasets}

