% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTrees.R
\name{FFTrees}
\alias{FFTrees}
\title{Creates a Fast and Frugal Trees (FFTrees) object.}
\usage{
FFTrees(formula = NULL, data = NULL, data.test = NULL, train.p = 1,
  rank.method = "m", hr.weight = 0.5, max.levels = 4,
  tree.definitions = NULL, verbose = FALSE, do.cart = TRUE,
  do.lr = TRUE, do.rf = TRUE, object = NULL)
}
\arguments{
\item{formula}{formula. A formula specifying a logical criterion as a function of 1 or more predictors.}

\item{data}{dataframe. A training dataset.}

\item{data.test}{dataframe. An optional testing dataset with the same structure as data.}

\item{train.p}{numeric. What percentage of the data to use for training. This only applies when data.test is not specified by the user.}

\item{rank.method}{character. How should cues be ranked during tree construction. "m" (for marginal) means that cues will only be ranked once with the entire training dataset. "c" (conditional) means that cues will be ranked after each level in the tree with the remaining unclassified training exemplars. This also means that the same cue can be used multiple times in the trees. However, the "c" method will take longer and may be prone to overfitting.}

\item{hr.weight}{numeric. A number between 0 and 1 indicating how much weight to give to maximizing hits versus minimizing false alarms.}

\item{max.levels}{integer. The maximum number of levels considered for the trees. Default is 5.}

\item{tree.definitions}{dataframe. An optional hard-coded definition of trees. See details.}

\item{verbose}{logical. Should progress reports be printed? Can be helpful for diagnosis when the function is running slowly...}

\item{do.cart, do.lr, do.rf}{logical. Should alternative algorithms be created for comparison? cart = regression trees, lr = logistic regression, rf = random forests.}

\item{object}{An optional existing FFTrees object (do not specify by hand)}
}
\value{
A list, see details
}
\description{
This is the workhorse function for the FFTrees package.
}
\details{
Here are the main elements of the output:
\describe{
  \item{cue.accuracies}{Marginal accuracies of each cue given a threshold that maximizes hr - far.}
  \item{tree.definitions}{Definitions of each tree created by \code{FFTrees}. Each row corresponds to one tree. Different levels within a tree are separated by semi-colons.}
  \item{tree.stats}{Tree definitions and classification statistics. Training and test data are stored separately}
  \item{level.stats}{Cumulative classification statistics at each tree level. Training and test data are stored separately}
  \item{decision}{Classification decisions for each case (rows) and each tree (columns). Training and test data are stored separately.}
  \item{levelout}{The level at which each case is classified in each tree. Rows correspond to cases and columns correspond to trees. Training and test data are stored separately.}
  \item{auc}{Area under the curve statistics}
  \item{lr}{Logistic regression model (without regularisation) and classification statistics.}
  \item{cart}{CART model and classification statistics}
  \item{rf}{Random Forest model and classification statistics}
  }
}

