% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fftrees_apply}
\alias{fftrees_apply}
\title{Applies a fast-and-frugal tree to a dataset and generates several accuracy statistics}
\usage{
fftrees_apply(x, allNA.pred = FALSE)
}
\arguments{
\item{x}{FFTrees.}

\item{allNA.pred}{logical. What should be predicted if all cue values in tree are NA? Default is FALSE}
}
\value{
A list of length 4 containing
}
\description{
Applies a fast-and-frugal tree to a dataset and generates several accuracy statistics
}
