% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_apply.R
\name{fftrees_apply}
\alias{fftrees_apply}
\title{Applies a FFT to a dataset and generates accuracy statistics.}
\usage{
fftrees_apply(x, mydata = NULL, newdata = NULL, allNA.pred = FALSE)
}
\arguments{
\item{x}{An FFT to apply (as an \code{FFTrees} object).}

\item{mydata}{dataframe.}

\item{newdata}{dataframe.}

\item{allNA.pred}{logical. What should be predicted if all cue values in tree are NA? Default is \code{FALSE}.}
}
\value{
A list of length 4 containing FFTs.
}
\description{
\code{fftrees_apply} applies a fast-and-frugal tree (FFT, as an \code{FFTrees} object)
to a dataset and generates corresponding accuracy statistics.
}
\seealso{
\code{\link{plot.FFTrees}} for plotting FFTs;
\code{\link{FFTrees}} for creating FFTs from and applying them to data.
}
\keyword{internal}
