% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_create.R
\name{fftrees_create}
\alias{fftrees_create}
\title{Create an object of class \code{FFTrees}}
\usage{
fftrees_create(
  data = NULL,
  formula = NULL,
  algorithm = NULL,
  goal = NULL,
  goal.chase = NULL,
  goal.threshold = NULL,
  numthresh.method = NULL,
  numthresh.n = NULL,
  sens.w = NULL,
  max.levels = NULL,
  cost.outcomes = NULL,
  cost.cues = NULL,
  stopping.rule = NULL,
  stopping.par = NULL,
  decision.labels = NULL,
  main = NULL,
  my.tree = NULL,
  data.test = NULL,
  repeat.cues = NULL,
  do.lr = TRUE,
  do.svm = TRUE,
  do.cart = TRUE,
  do.rf = TRUE,
  do.comp = TRUE,
  quiet = NULL
)
}
\arguments{
\item{data}{Training data (as data frame).}

\item{formula}{A formula (with a binary criterion variable).}

\item{algorithm}{string.}

\item{goal}{string.}

\item{goal.chase}{string.}

\item{goal.threshold}{string.}

\item{numthresh.method}{string.}

\item{numthresh.n}{integer.}

\item{sens.w}{numeric.}

\item{max.levels}{integer.}

\item{cost.outcomes}{list.}

\item{cost.cues}{list.}

\item{stopping.rule}{string.}

\item{stopping.par}{numeric.}

\item{decision.labels}{string.}

\item{main}{string.}

\item{my.tree}{string.}

\item{data.test}{dataframe.}

\item{repeat.cues}{logical.}

\item{do.lr}{logical.}

\item{do.svm}{logical.}

\item{do.cart}{logical.}

\item{do.rf}{logical.}

\item{do.comp}{logical.}

\item{quiet}{logical}
}
\value{
An \code{FFTrees} object.
}
\description{
\code{fftrees_create} creates an \code{FFTrees} object.

\code{fftrees_create} is called internally by the main \code{\link{FFTrees}} function.
}
\seealso{
\code{\link{fftrees_define}} for defining FFTs;
\code{\link{FFTrees}} for creating FFTs from and applying them to data.
}
\keyword{internal}
