% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_cuerank_function.R
\name{fftrees_cuerank}
\alias{fftrees_cuerank}
\title{Calculates thresholds that maximize a statistic (goal) for cues.}
\usage{
fftrees_cuerank(x = NULL, data_train = NULL, rounding = NULL)
}
\arguments{
\item{x}{FFTrees.}

\item{data_train}{dataframe.}
}
\value{
A dataframe containing thresholds and marginal classification statistics for each cue
}
\description{
Calculates thresholds that maximize a statistic (goal) for cues.
}
\examples{


\dontrun{
 # What are the best thresholds for each cue in the mushrooms dataset?
 mushrooms.cues <- cuerank(formula = poisonous ~.,
                           data = mushrooms)
}


}
