% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_cuerank.R
\name{fftrees_cuerank}
\alias{fftrees_cuerank}
\title{Calculate thresholds that optimize some statistic (goal) for cues in data}
\usage{
fftrees_cuerank(x = NULL, newdata = NULL, data = "train", rounding = NULL)
}
\arguments{
\item{x}{An \code{FFTrees} object.}

\item{newdata}{dataframe.}

\item{data}{dataframe.}

\item{rounding}{integer.}
}
\value{
A data frame containing thresholds and marginal classification statistics for each cue
}
\description{
Calculate thresholds that optimize some statistic (goal) for cues in data
}
