% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_cuerank.R
\name{fftrees_cuerank}
\alias{fftrees_cuerank}
\title{Calculates thresholds that maximize a statistic (goal) for cues.}
\usage{
fftrees_cuerank(x = NULL, newdata = NULL, data = "train", rounding = NULL)
}
\arguments{
\item{x}{FFTrees.}

\item{newdata}{dataframe.}

\item{data}{dataframe.}

\item{rounding}{integer.}
}
\value{
A dataframe containing thresholds and marginal classification statistics for each cue
}
\description{
Calculates thresholds that maximize a statistic (goal) for cues.
}
