% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_define.R
\name{fftrees_define}
\alias{fftrees_define}
\title{Create definitions of FFTs.}
\usage{
fftrees_define(x, object = NULL)
}
\arguments{
\item{x}{FFTrees.}

\item{object}{FFTrees.}
}
\value{
An \code{FFTrees} object with tree definitions.
}
\description{
\code{fftrees_define} creates definitions of fast-and-frugal trees
(FFTs, as an \code{FFTrees} object).
}
\keyword{internal}
