% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_define.R
\name{fftrees_define}
\alias{fftrees_define}
\title{Create definitions of FFTrees}
\usage{
fftrees_define(x, object = NULL)
}
\arguments{
\item{x}{FFTrees.}

\item{object}{FFTrees.}
}
\value{
An FFTrees object with tree definitions
}
\description{
Create definitions of FFTrees
}
\keyword{internal}
