% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftrees_wordstofftrees.R
\name{fftrees_wordstofftrees}
\alias{fftrees_wordstofftrees}
\title{Converts text describing an FFT into an FFT definition.}
\usage{
fftrees_wordstofftrees(x, my.tree)
}
\arguments{
\item{x}{FFTrees.}

\item{my.tree}{string. A string defining an FFT}
}
\description{
Converts text describing an FFT into an FFT definition.
}
