% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartcost_doc.R
\docType{data}
\name{heart.cost}
\alias{heart.cost}
\title{Cue costs for the heartdisease dataa}
\format{
A data frame containing 153 rows and 14 columns
\describe{
  \item{cue}{The name of the cue}
  \item{cost}{The cost of the cue}



  ...
}
}
\source{
https://archive.ics.uci.edu/ml/machine-learning-databases/heart-disease/costs/
}
\usage{
heart.cost
}
\description{
Cue costs for the heartdisease dataa
}
\keyword{datasets}
