% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inwords.R
\name{inwords}
\alias{inwords}
\title{Provide a verbal description of an FFT}
\usage{
inwords(x, tree = 1)
}
\arguments{
\item{x}{An \code{FFTrees} object.}

\item{tree}{The tree to display (as numeric).}
}
\value{
A verbal description of an FFT (as a string).
}
\description{
\code{inwords} generates and provides a verbal description
of a fast-and-frugal tree (FFT) from an \code{FFTrees} object.
}
\seealso{
\code{\link{print.FFTrees}} for printing FFTs;
\code{\link{plot.FFTrees}} for plotting FFTs;
\code{\link{summary.FFTrees}} for summarizing FFTs;
\code{\link{FFTrees}} for creating FFTs from and applying them to data.
}
