% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inwords.R
\name{inwords}
\alias{inwords}
\title{Display a verbal description of a tree in an FFTrees object}
\usage{
inwords(x, tree = 1)
}
\arguments{
\item{x}{FFTrees.}

\item{tree}{numeric. The tree to display}
}
\value{
A string describing an FFT
}
\description{
Display a verbal description of a tree in an FFTrees object
}
