% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irisv_doc.R
\docType{data}
\name{iris.v}
\alias{iris.v}
\title{Iris data set}
\format{A data frame containing 150 rows and 4 columns}
\source{
https://archive.ics.uci.edu/ml/datasets/Iris
}
\usage{
iris.v
}
\description{
Iris data set
}
\keyword{datasets}
