% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{iris.v}
\alias{iris.v}
\title{Iris data}
\format{
A data frame containing 150 rows and 4 columns.

\describe{
  \item{sep.len}{sepal length in cm}
  \item{sep.wid}{sepal width in cm}
  \item{pet.len}{petal length in cm}
  \item{pet.wid}{petal width in cm}
  \item{virginica}{\emph{Criterion}: Does an iris belong to the class "virginica"?

  Values: \code{TRUE} vs. \code{FALSE} (33.33\% vs.66.67\%).}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Iris}
}
\usage{
iris.v
}
\description{
A famous dataset from R.A. Fisher (1936) simplified to predict only the virginica class (i.e., as a binary classification problem).
}
\details{
To improve usability, we made the following changes:

\itemize{

 \item{The criterion was binarized from a factor variable with three levels
 (\code{Iris-setosa}, \code{Iris-versicolor}, \code{Iris-virginica}),
 into a logical variable (i.e., \code{TRUE} for all instances of \code{Iris-virginica}
 and \code{FALSE} for the two other levels).}

 }

 Other than that, the data remains consistent with the original dataset.
}
\references{
Fisher, R.A. (1936): The use of multiple measurements in taxonomic problems.
Annual Eugenics, 7, Part II, pp. 179--188.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
