% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mushrooms_doc.R
\docType{data}
\name{mushrooms}
\alias{mushrooms}
\title{Mushrooms dataset}
\format{A data frame containing 8,124 rows and 23 columns (see http://archive.ics.uci.edu/ml/machine-learning-databases/mushroom/agaricus-lepiota.names for column descriptions)
\describe{
  \item{poisonous}{numeric}
  \item{cshape}{character}
  \item{csurface}{character}
  \item{ccolor}{character}
 \item{bruises}{character}
 \item{odor}{numeric}
 \item{gattach}{character}
 \item{gspace}{characte}
 \item{gsize}{character}
 \item{gcolor}{character}
 \item{sshape}{character}
 \item{sroot}{character}
 \item{ssaring}{character}
 \item{ssbring}{character}
 \item{scaring}{character}
 \item{scbring}{character}
 \item{vtype}{character}
 \item{vcolor}{character}
 \item{ringnum}{character}
 \item{ringtype}{character}
 \item{sporepc}{character}
 \item{population}{character}
 \item{habitat}{character}
  ...
}}
\source{
http://archive.ics.uci.edu/ml/datasets/Mushroom
}
\usage{
mushrooms
}
\description{
Mushrooms dataset
}
\keyword{datasets}
