% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{mushrooms}
\alias{mushrooms}
\title{Mushrooms data}
\format{
A data frame containing 8,124 rows and 23 columns.

See \url{http://archive.ics.uci.edu/ml/machine-learning-databases/mushroom/agaricus-lepiota.names} for column descriptions.

\describe{
 \item{poisonous}{logical criterion variable}
 \item{cshape}{character}
 \item{csurface}{character}
 \item{ccolor}{character}
 \item{bruises}{character}
 \item{odor}{numeric}
 \item{gattach}{character}
 \item{gspace}{characte}
 \item{gsize}{character}
 \item{gcolor}{character}
 \item{sshape}{character}
 \item{sroot}{character}
 \item{ssaring}{character}
 \item{ssbring}{character}
 \item{scaring}{character}
 \item{scbring}{character}
 \item{vtype}{character}
 \item{vcolor}{character}
 \item{ringnum}{character}
 \item{ringtype}{character}
 \item{sporepc}{character}
 \item{population}{character}
 \item{habitat}{character}

  ...
}
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Mushroom}
}
\usage{
mushrooms
}
\description{
Data describing poisonous vs. non-poisonous mushrooms.
}
\details{
This dataset includes descriptions of hypothetical samples corresponding to 23 species of gilled mushrooms
in the Agaricus and Lepiota Family. Each species is classified as \code{poisonous} (True or False).
The Guide clearly states that there is no simple rule for determining the edibility of a mushroom;
no rule like ``leaflets three, let it be'' for Poisonous Oak and Ivy.
}
\references{
Mushroom records drawn from The Audubon Society Field Guide to North American Mushrooms (1981).
G.H. Lincoff (Pres.), New York: A.A. Knopf.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
