% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFFForest.R
\name{plot.FFForest}
\alias{plot.FFForest}
\title{Creates a network plot. Code taken from Dirk Wulff (www.dirkwulff.org)}
\usage{
\method{plot}{FFForest}(x, node.cex.lim = c(1, 10), line.cex.lim = c(0.3,
  5), mincon = 0, lo = "kk", palette = NULL, ...)
}
\arguments{
\item{x}{FFForest. An FFForest object created from FFForest()}

\item{node.cex.lim}{numeric. Nodesize adjustment}

\item{line.cex.lim}{numeric. Edgesize adjustment}

\item{mincon}{integer. Minimum connection cutoff}

\item{lo}{string. The layout of the network plot. Either 'kk' (Kamada-Kawai, the default), 'dh' (Davidson-Harel) or 'fr' (Fruchterman-Reingold)}

\item{palette}{string. A string vector of colors}

\item{...}{currently ignored}
}
\description{
Creates a network plot. Code taken from Dirk Wulff (www.dirkwulff.org)
}
