% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFFTrees_function.R
\name{plot.FFTrees}
\alias{plot.FFTrees}
\title{Plots an FFTrees object.}
\usage{
\method{plot}{FFTrees}(
  x = NULL,
  data = "train",
  what = "tree",
  tree = 1,
  main = NULL,
  hlines = TRUE,
  cue.labels = NULL,
  decision.labels = NULL,
  cue.cex = NULL,
  threshold.cex = NULL,
  decision.cex = 1,
  comp = TRUE,
  stats = TRUE,
  show.header = NULL,
  show.tree = NULL,
  show.confusion = NULL,
  show.levels = NULL,
  show.roc = NULL,
  show.icons = NULL,
  show.iconguide = NULL,
  label.tree = NULL,
  label.performance = NULL,
  n.per.icon = NULL,
  which.tree = NULL,
  level.type = "bar",
  decision.names = NULL,
  ...
)
}
\arguments{
\item{x}{A FFTrees object created from \code{"FFTrees()"}}

\item{data}{One of two strings 'train' or 'test'. In this case, the corresponding dataset in the x object will be used.}

\item{what}{string. What should be plotted? \code{'tree'} (the default) shows one tree (specified by \code{'tree'}). \code{'cues'} shows the marginal accuracy of cues in an ROC space, \code{"roc"} shows an roc curve of the tree(s)}

\item{tree}{integer. An integer indicating which tree to plot (only valid when the tree argument is non-empty). To plot the best training (or test) tree with respect to the \code{goal} specified during FFT construction, use "best.train" or "best.test"}

\item{main}{character. The main plot label.}

\item{hlines}{logical. Should horizontal panel separation lines be shown?}

\item{cue.labels}{character. An optional string of labels for the cues / nodes.}

\item{decision.labels}{character. A string vector of length 2 indicating the content-specific name for noise and signal cases.}

\item{cue.cex}{numeric. The size of the cue labels.}

\item{threshold.cex}{numeric. The size of the threshold labels.}

\item{decision.cex}{numeric. The size of the decision labels.}

\item{comp}{logical. Should the performance of competitive algorithms (e.g.; logistic regression, random forests etc.) be shown in the ROC plot (if available?)}

\item{stats}{logical. Should statistical information be plotted? If \code{FALSE}, then only the tree (without any reference to statistics) will be plotted.}

\item{show.header, show.tree, show.confusion, show.levels, show.roc, show.icons, show.iconguide}{logical. Logical arguments indicating which specific elements of the plot to show.}

\item{label.tree, label.performance}{string. Optional arguments to define lables for the tree and performance section(s).}

\item{n.per.icon}{Number of cases per icon}

\item{which.tree}{deprecated argument, only for backwards compatibility, use \code{"tree"} instead.}

\item{level.type}{string. How should bottom levels be drawn? Can be \code{"bar"} or \code{"line"}}

\item{decision.names}{deprecated arguments.}

\item{...}{Currently ignored.}
}
\value{
A plot vizualizing an FFT
}
\description{
Plots an FFTrees object created by the FFTrees() function.
}
\examples{

# Create FFTrees of the heart disease data
heart.fft <- FFTrees(
  formula = diagnosis ~ .,
  data = heartdisease
)

# Visualise the tree
plot(heart.fft,
  main = "Heart Disease Diagnosis",
  decision.labels = c("Absent", "Present")
)


# See the vignette for more details
vignette("FFTrees_plot", package = "FFTrees")

}
