% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFFTrees_function.R
\name{predict.FFTrees}
\alias{predict.FFTrees}
\title{Predict new data with an FFTrees object}
\usage{
\method{predict}{FFTrees}(object = NULL, data = NULL, tree = "best.train",
  ...)
}
\arguments{
\item{object}{An FFTrees object created from the FFTrees() function.}

\item{data}{A dataframe of test data}

\item{tree}{Which tree in the FFTrees object should be used? Can be an integer or "best.train" (the default) to use the tree with the best training statistics.}

\item{...}{Additional arguments passed on to \code{predict()}}
}
\value{
A logical vector of predictions
}
\description{
Predict new data with an FFTrees object
}
\examples{


  # Create an FFTrees object from 200 cases from thethe breastcancer dataset

  breastcancer.fft <- FFTrees(formula = diagnosis ~.,
                              data = breastcancer[1:300,])

 # Predict results for remaining data
  predict(breastcancer.fft,
  data = breastcancer[301:nrow(breastcancer),])




}

