% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFFTrees_function.R
\name{predict.FFTrees}
\alias{predict.FFTrees}
\title{Applies an existing FFTrees object to a new dataset}
\usage{
\method{predict}{FFTrees}(object = NULL, data.test = NULL, ...)
}
\arguments{
\item{object}{(M) An FFTrees object created from the FFTrees() function.}

\item{data.test}{(M) A dataframe of test data}

\item{...}{Additional arguments passed on to predict()}
}
\value{
An FFTrees object
}
\description{
Applies an existing FFTrees object to a new dataset
}

