% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFFTrees_function.R
\name{print.FFTrees}
\alias{print.FFTrees}
\title{Prints summary information from an FFTrees x}
\usage{
\method{print}{FFTrees}(x = NULL, ...)
}
\arguments{
\item{x}{FFTrees. A FFTrees x created from FFTrees()}

\item{...}{additional arguments passed to print.}
}
\description{
Printing function for an FFTrees x
}
