% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcues_function.R
\name{showcues}
\alias{showcues}
\title{Visualizes cue accuracies from an FFTrees object in a ROC space}
\usage{
showcues(x = NULL, data = "train", main = NULL, top = 5,
  palette = c("#0C5BB0CC", "#EE0011CC", "#15983DCC", "#EC579ACC", "#FA6B09CC",
  "#149BEDCC", "#A1C720CC", "#FEC10BCC", "#16A08CCC", "#9A703ECC"))
}
\arguments{
\item{x}{An FFTrees object}

\item{data}{A string indicating whether or not to show training ("train") or testing ("test") cue accuracies}

\item{main}{Main plot description}

\item{top}{An integer indicating how many of the top cues to highlight}

\item{palette}{An optional vector of colors}
}
\description{
Visualizes cue accuracies from an FFTrees object in a ROC space
}
