% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcues_function.R
\name{showcues}
\alias{showcues}
\title{Visualize cue accuracies (as points in ROC space)}
\usage{
showcues(x = NULL, data = "train", cue.accuracies = NULL, main = NULL, top = 5)
}
\arguments{
\item{x}{An \code{FFTrees} object created by the \code{\link{FFTrees}} function.}

\item{data}{A string indicating whether to show cue accuracies in training (\code{"train"}) or testing (\code{"test"}).}

\item{cue.accuracies}{An optional data frame specifying cue accuracies directly (without specifying an \code{FFTrees} object x).}

\item{main}{Main plot title (as character string).}

\item{top}{How many of the top cues should be highlighted (as an integer)?}
}
\value{
A plot showing cue accuracies (of an \code{FFTrees} object) (as points in ROC space).
}
\description{
\code{showcues} plots the cue accuracies of an \code{FFTrees} object
created by the \code{\link{FFTrees}} function (as points in ROC space).

\code{showcues} is called when the main \code{\link{plot.FFTrees}} function is set to \code{what = "cues"}.
}
\examples{
# Create fast-and-frugal trees (FFTs) for heart disease:
heart.fft <- FFTrees(formula = diagnosis ~ .,
                     data = heart.train,
                     data.test = heart.test,
                     main = "Heart Disease",
                     decision.labels = c("Healthy", "Diseased")
                     )

# Show cue accuracies (in ROC space):
showcues(heart.fft,
         main = "Cue accuracies for predicting heart disease")

}
\seealso{
\code{\link{plot.FFTrees}} for plotting FFTs;
\code{\link{print.FFTrees}} for printing summary information of FFTs;
\code{\link{FFTrees}} for creating FFTs from and applying them to data.

Other plot functions: 
\code{\link{plot.FFTrees}()}
}
\concept{plot functions}
