% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFFTrees.R
\name{simFFTrees}
\alias{simFFTrees}
\title{Generates and summarises many FFTrees objects from multiple training stes}
\usage{
simFFTrees(formula = NULL, data = NULL, max.levels = 5, sim = 10,
  train.p = 0.5, rank.method = "m", hr.weight = 0.5, verbose = TRUE,
  cpus = 1)
}
\arguments{
\item{formula}{A formula specifying a binary criterion as a function of multiple variables}

\item{data}{A dataframe containing variables in formula}

\item{max.levels}{A number indicating the maximum number of levels considered for the tree.}

\item{sim}{integer. Number of simulations to perform.}

\item{train.p}{A number between 0 and 1 indicating what percentage of the data to use for training. This only applies when data.test is not specified by the user.}

\item{rank.method}{rank.method A string indicating how to rank cues during tree construction. "m" (for marginal) means that cues will only be ranked once with the entire training dataset. "c" (conditional) means that cues will be ranked after each level in the tree with the remaining unclassified training exemplars. This also means that the same cue can be used multiple times in the trees. However, the "c" method will take longer and may be prone to overfitting.}

\item{hr.weight}{A number between 0 and 1 indicating how much weight to give to maximizing hits versus minimizing false alarms.}

\item{verbose}{A logical value indicating whether or not to print progress reports.}

\item{cpus}{integer. Number of cpus to use (any value larger than 1 will initiate parallel calculations in snowfall)}
}
\value{
A dataframe containing best thresholds and marginal classification statistics for each cue
}
\description{
Generates and summarises many FFTrees objects from multiple training stes
}
\examples{

train.5m <- simFFTrees(formula = diagnosis ~.,
                       cpus = 1)



}

