% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{sonar}
\alias{sonar}
\title{Sonar data}
\format{
A data frame containing 208 rows and 60 columns.


\describe{
  \item{V1}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V2}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V3}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V4}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V5}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V6}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V7}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V8}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V9}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V10}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V11}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V12}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V13}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V14}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V15}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V16}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V17}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V18}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V19}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V20}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V21}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V22}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V23}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V24}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V25}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V26}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V27}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V28}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V29}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V30}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V31}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V32}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V33}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V34}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V35}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V36}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V37}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V38}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V39}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V40}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V41}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V42}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V43}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V44}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V45}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V46}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V47}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V48}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V49}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V50}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V51}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V52}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V53}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V54}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V55}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V56}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V57}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V58}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V59}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{V60}{Number in the range 0.0 to 1.0 that represents the energy within a particular frequency band, integrated over a certain period of time.}
  \item{mine.crit}{\emph{Criterion}: Did a sonar signal bounce off a metal cylinder (or a rock)?

  Values: \code{TRUE} (metal cylinder) vs. \code{FALSE} (rock) (53.37\% vs.\ 46.63\%).}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Connectionist+Bench+(Sonar,+Mines+vs.+Rocks)}
}
\usage{
sonar
}
\description{
The file contains patterns of sonar signals bounced off a metal cylinder or bounced off a roughly cylindrical rock at various angles and under various conditions.
The transmitted sonar signal is a frequency-modulated chirp, rising in frequency.
}
\details{
We made the following enhancements to the original data for improved usability:

\itemize{
 \item{The binary factor \emph{criterion} variable with exclusive "m" and "r" values was converted to a logical \code{TRUE/FALSE} vector.}
 }

 Other than that, the data remains consistent with the original dataset.
}
\references{
Gorman, R. P., and Sejnowski, T. J. (1988). "Analysis of Hidden Units in a Layered Network Trained to Classify Sonar Targets" in Neural Networks, Vol. 1, pp. 75-89.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{titanic}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
