% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_doc.R
\docType{data}
\name{sonar}
\alias{sonar}
\title{Sonar data set}
\format{A data frame containing 208 rows and 60 columns}
\source{
https://archive.ics.uci.edu/ml/datasets/Connectionist+Bench+(Sonar,+Mines+vs.+Rocks)
}
\usage{
sonar
}
\description{
Sonar data set
}
\keyword{datasets}

