% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFFTrees_function.R
\name{summary.FFTrees}
\alias{summary.FFTrees}
\title{Returns summary information about an FFTrees x}
\usage{
\method{summary}{FFTrees}(x, tree = 1, ...)
}
\arguments{
\item{x}{An FFTrees x}

\item{tree}{integer. The tree to summarise}

\item{...}{additional arguments (currently ignored)}
}
\description{
Returns summary information about an FFTrees x
}
