% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFFTrees_function.R
\name{summary.FFTrees}
\alias{summary.FFTrees}
\title{Returns a summary of an fft object}
\usage{
\method{summary}{FFTrees}(object, ...)
}
\arguments{
\item{object}{An FFTrees object}

\item{...}{Additional arguments}
}
\description{
Returns a summary of an fft object
}

