% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic survival data}
\format{
A data frame containing 2,201 rows and 4 columns.

\describe{
  \item{class}{Factor - Class (first, second, third, or crew)}
  \item{age}{Factor - Age group (child or adult)}
  \item{sex}{Factor - Sex (male or female)}
  \item{survived}{Factor - Whether the passenger survived (1) or not (0)}

  ...
}
}
\source{
\url{https://www.encyclopedia-titanica.org}
}
\usage{
titanic
}
\description{
Data indicating who survived on the Titanic.
}
\details{
See \code{\link{Titanic}} of the R \strong{datasets} package for details and
the same data (in a 4-dimensional \code{table}).
}
\references{
Dawson, Robert J. MacG. (1995), The ‘Unusual Episode’ Data Revisited. Journal of Statistics Education, 3.
doi: 10.1080/10691898.1995.11910499.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{voting}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
