% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic_doc.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic dataset}
\format{A data frame containing 2,201 rows and 4 columns
\describe{
  \item{class}{Factor - Class (first, second, third, or crew)}
  \item{age}{Factor - Age group (child or adult)}
  \item{sex}{Factor - Sex (male or female)}
  \item{survived}{Factor - Whether the passenger survived (1) or not (0)}

  ...
}}
\source{
https://www.amstat.org/publications/jse/datasets/titanic.dat.txt
}
\usage{
titanic
}
\description{
A dataset indicating who survived on the Titanic
}
\keyword{datasets}

