% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFTreesdata_doc.R
\docType{data}
\name{voting}
\alias{voting}
\title{Voting data}
\format{
A data frame containing 435 rows and 16 columns.
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Congressional+Voting+Records}
}
\usage{
voting
}
\description{
A dataset of votes for each of the U.S. House of Representatives Congressmen on the 16 key votes identified by the CQA.
}
\details{
The CQA lists nine different types of votes: voted for, paired for, and announced for (these three simplified to yea),
voted against, paired against, and announced against (these three simplified to nay),
voted present, voted present to avoid conflict of interest, and did not vote or otherwise make a position known
(these three simplified to an unknown disposition).

The binary criterion variable used here is \code{party.crit}.
}
\references{
Congressional Quarterly Almanac, 98th Congress, 2nd session 1984, Volume XL: Congressional Quarterly Inc.
Washington, D.C., 1985.
}
\seealso{
Other datasets: 
\code{\link{blood}},
\code{\link{breastcancer}},
\code{\link{car}},
\code{\link{contraceptive}},
\code{\link{creditapproval}},
\code{\link{fertility}},
\code{\link{forestfires}},
\code{\link{heart.cost}},
\code{\link{heart.test}},
\code{\link{heart.train}},
\code{\link{heartdisease}},
\code{\link{iris.v}},
\code{\link{mushrooms}},
\code{\link{sonar}},
\code{\link{titanic}},
\code{\link{wine}}
}
\concept{datasets}
\keyword{datasets}
