% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_doc.R
\docType{data}
\name{voting}
\alias{voting}
\title{Voting data set}
\format{A data frame containing 435 rows and 16 columns}
\source{
https://archive.ics.uci.edu/ml/datasets/Congressional+Voting+Records
}
\usage{
voting
}
\description{
Voting data set
}
\keyword{datasets}

