% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine_doc.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine tasting dataset}
\format{A data frame containing 6497 rows and 13 columns}
\source{
http://archive.ics.uci.edu/ml/datasets/Wine+Quality
}
\usage{
wine
}
\description{
Chemical and tasting data from wines in North Portugal
}
\keyword{datasets}

