% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordstoFFT.R
\name{wordstoFFT}
\alias{wordstoFFT}
\title{Converts text describing an FFT into an FFT definition.}
\usage{
wordstoFFT(input, cue.names, decision.labels = NULL)
}
\arguments{
\item{input}{string. A string describing an FFT in words (see examples)}

\item{cue.names}{string. A vector of cue names}

\item{decision.labels}{string. A vector of decision labels}
}
\description{
Converts text describing an FFT into an FFT definition.
}
\examples{

<<<<<<< HEAD
my.tree.def <- wordstoFFT(input = "If age > 55, predict True.
                                   If cp = [a,b,np], predict False, otherwise, predict True",
=======
my.tree.def <- wordstoFFT(input = "If age > 55, predict True. If cp = [a,b,np], predict False, otherwise, predict True",
>>>>>>> 900075e9c02f7358588b27289bab19cae450ed66
                          cue.names = names(heartdisease))


}
